@@ -1,138 +1,101 @@
 # shadowstep/exceptions/shadowstep_exceptions.py
+from __future__ import annotations
 
+import datetime
 import traceback
-import typing
-from typing import Optional, Union, Tuple, List, Dict
+from collections.abc import Sequence
+from typing import Any
 
-import numpy as np
-from PIL import Image
-from appium.webdriver import WebElement
-from selenium.common import WebDriverException
-from selenium.types import WaitExcTypes
+from appium.webdriver.webdriver import WebDriver
+from selenium.common import NoSuchElementException, TimeoutException, WebDriverException
 
 
 class ShadowstepException(WebDriverException):
     """Raised when driver is not specified and cannot be located."""
 
     def __init__(
-            self, msg: typing.Optional[str] = None, screen: typing.Optional[str] = None,
-            stacktrace: typing.Optional[typing.Sequence[str]] = None
+            self,
+            msg: str | None = None,
+            screen: str | None = None,
+            stacktrace: Sequence[str] | None = None
     ) -> None:
         super().__init__(msg, screen, stacktrace)
 
 
 class ShadowstepElementError(ShadowstepException):
-    def __init__(self, message: str = '', original_exception: Optional[Exception] = None):
+    def __init__(self,
+                 message: str | None = None,
+                 original_exception: Exception | None = None):
         super().__init__(message)
         self.original_exception = original_exception
         self.traceback = traceback.format_exc()
 
 
-class ShadowstepGetElementError(ShadowstepElementError):
+class ShadowstepNoSuchElementError(NoSuchElementException):
     def __init__(self,
-                 message='Failed to get element',
-                 locator=None,
-                 contains=None,
-                 timeout: float = None,
-                 poll_frequency: float = 0.5,
-                 ignored_exceptions: Optional[WaitExcTypes] = None,
-                 original_exception: Optional[Exception] = None):
-        message = f"{message}. Args: {locator=}, {contains=}" if "Args:" not in message else message
-        super().__init__(message, original_exception)
+                 msg: str | None = None,
+                 screen: str | None = None,
+                 stacktrace: list[Any] | None = None,
+                 locator: Any = None):
+        super().__init__(msg, screen, stacktrace)
         self.locator = locator
-        self.contains = contains
-        self.timeout = timeout
-        self.poll_frequency = poll_frequency
-        self.ignored_exceptions = ignored_exceptions
+        self.msg = msg
+        self.screen = screen
+        self.stacktrace = stacktrace
+
+    def __str__(self):
+        return f"ShadowstepNoSuchElementError: Locator: {self.locator} \n Message: {self.msg} \n Stacktrace: {self.stacktrace}"
 
 
-class ShadowstepImageProcessingError(ShadowstepException):
-    def __init__(self, message: str = '',
-                 image: Union[bytes, str] = None,
-                 full_image: Union[bytes, str] = None,
-                 threshold: Optional[float] = None,
-                 original_exception: Optional[Exception] = None):
-        super().__init__(message)
-        self.image = image
-        self.full_image = full_image
-        self.threshold = threshold
-        self.original_exception = original_exception
-        self.traceback = traceback.format_exc()
+class ShadowstepTimeoutException(TimeoutException):
+    """Custom timeout exception with additional context."""
 
-
-class ShadowstepTextRecognitionError(ShadowstepException):
     def __init__(self,
-                 message: str = 'Failed to recognize text',
-                 text: Optional[str] = None,
-                 language: Optional[str] = None,
-                 image: Union[bytes, str] = None,
-                 contains: Optional[bool] = None,
-                 ocr: Optional[bool] = None,
-                 original_exception: Optional[Exception] = None):
-        message = f"{message}. Args: {text=}, {language=}, {ocr=}, {contains=}" if "Args:" not in message else message
-        super().__init__(message)
-        self.text = text
-        self.language = language
-        self.image = image
-        self.ocr = ocr
-        self.contains = contains
-        self.original_exception = original_exception
-        self.traceback = traceback.format_exc()
-
-
-class ShadowstepSwipeError(ShadowstepException):
-    def __init__(self,
-                 message: str = 'Swipe action failed',
-                 start_position=None,
-                 end_position=None,
-                 direction=None,
-                 distance=None,
-                 duration=None,
-                 original_exception: Optional[Exception] = None):
-        message = f"{message}. Args: {start_position=}, {end_position=}, {direction=}, {distance=}, {duration=}" if "Args:" not in message else message
-        super().__init__(message)
-        self.start_position = start_position
-        self.end_position = end_position
-        self.direction = direction
-        self.distance = distance
-        self.duration = duration
-        self.original_exception = original_exception
-        self.traceback = traceback.format_exc()
-
-
-class ShadowstepTimeoutError(ShadowstepException):
-    def __init__(self,
-                 message: str = 'Timeout exceeded',
-                 locator: Optional[Union[str, Dict]] = None,
-                 image: Optional[Union[str, bytes, np.ndarray, Image.Image]] = None,
-                 timeout: float = 10.0,
-                 contains: Optional[bool] = True,
-                 original_exception: Optional[Exception] = None):
-        message = f"{message}. Args: {locator=}, {image=}, {timeout=}, {contains=}" if "Args:" not in message else message
-        super().__init__(message)
+                 msg: str | None = None,
+                 screen: str | None = None,
+                 stacktrace: list[Any] | None = None,
+                 locator: Any = None,
+                 driver: WebDriver | None = None):
+        super().__init__(msg, screen, stacktrace)
         self.locator = locator
-        self.image = image
-        self.timeout = timeout
-        self.contains = contains
-        self.original_exception = original_exception
-        self.traceback = traceback.format_exc()
+        self.driver = driver
+        self.timestamp = datetime.datetime.now().strftime("%Y-%m-%d_%H-%M-%S")
+
+    def __str__(self):
+        return (f"ShadowstepTimeoutException\n"
+                f"Timestamp: {self.timestamp}\n"
+                f"Message: {self.msg}\n"
+                f"Locator: {self.locator}\n"
+                f"Current URL: {self.driver.current_url if self.driver else 'N/A'}\n"
+                f"Stacktrace:\n{''.join(self.stacktrace) if self.stacktrace else 'N/A'}")
 
 
-class ShadowstepTapError(ShadowstepException):
-    def __init__(self,
-                 message: str = 'Tap action failed',
-                 locator: Optional[Union[str, Dict]] = None,
-                 x: Optional[int] = None,
-                 y: Optional[int] = None,
-                 image: Optional[Union[str, bytes]] = None,
-                 duration: Optional[int] = None,
-                 original_exception: Optional[Exception] = None):
-        message = f"{message}. Args: {locator=}, {x=}, {y=}, {duration=}" if "Args:" not in message else message
-        super().__init__(message)
-        self.locator = locator
-        self.x = x
-        self.y = y
-        self.image = image
-        self.duration = duration
-        self.original_exception = original_exception
-        self.traceback = traceback.format_exc()
+class ShadowstepElementException(WebDriverException):
+    """Raised when driver is not specified and cannot be located."""
+
+    def __init__(
+            self, msg: str | None = None, screen: str | None = None,
+            stacktrace: Sequence[str] | None = None
+    ) -> None:
+        super().__init__(msg, screen, stacktrace)
+
+
+class ShadowstepLocatorConverterError(Exception):
+    """Base exception for locator conversion errors."""
+    pass
+
+
+class ShadowstepInvalidUiSelectorError(Exception):
+    """Raised when UiSelector string is malformed."""
+    pass
+
+
+class ShadowstepConversionError(ShadowstepLocatorConverterError):
+    """Raised when conversion between formats fails."""
+    pass
+
+
+class ShadowstepResolvingLocatorError(Exception):
+    """Raised when locator resolving is failed (used in shadowstep.element.dom)."""
+    pass
+
