@@ -1,21 +1,30 @@
-import inspect
+# shadowstep/mobile_commands.py
+from __future__ import annotations
+
 import logging
-from typing import Union, Dict, List
+from typing import TYPE_CHECKING, Any
 
-from selenium.common.exceptions import NoSuchDriverException, InvalidSessionIdException, StaleElementReferenceException
+from selenium.common.exceptions import (
+    InvalidSessionIdException,
+    NoSuchDriverException,
+    StaleElementReferenceException,
+)
 
-from shadowstep.exceptions.shadowstep_exceptions import ShadowstepException
-from shadowstep.utils.decorators import fail_safe
+if TYPE_CHECKING:
+    from shadowstep.shadowstep import Shadowstep
 from shadowstep.base import WebDriverSingleton
-
+from shadowstep.decorators.decorators import fail_safe
+from shadowstep.exceptions.shadowstep_exceptions import ShadowstepException
+from shadowstep.utils.utils import get_current_func_name
 
 
 class MobileCommands:
 
-    def __init__(self, shadowstep: 'Shadowstep'):
-        self.driver = None
-        self.logger = logging.getLogger(f"{__name__}.{self.__class__.__name__}")
+    def __init__(self, shadowstep: Shadowstep):
         self.shadowstep = shadowstep
+        self.driver: Any = None
+        self.logger = logging.getLogger(f"{__name__}.{self.__class__.__name__}")
+        
 
     @fail_safe(
         retries=3,
@@ -23,7 +32,7 @@ class MobileCommands:
         raise_exception=ShadowstepException,
         exceptions=(NoSuchDriverException, InvalidSessionIdException, StaleElementReferenceException)
     )
-    def activate_app(self, params: Union[Dict, List] = None) -> 'Shadowstep':
+    def activate_app(self, params: dict[str, Any] | list[Any]) -> MobileCommands:
         """
         https://github.com/appium/appium-uiautomator2-driver/blob/master/docs/android-mobile-gestures.md#mobile activateapp
         Execute mobile: activateApp command.
@@ -34,8 +43,8 @@ class MobileCommands:
         Returns:
             Shadowstep: Self for method chaining.
         """
-        self.logger.debug(f"{inspect.currentframe().f_code.co_name}")
-        self._execute('mobile: activateApp', params or {})
+        self.logger.debug(f"{get_current_func_name()}")
+        self._execute("mobile: activateApp", params)
         return self
 
 
@@ -45,7 +54,7 @@ class MobileCommands:
         raise_exception=ShadowstepException,
         exceptions=(NoSuchDriverException, InvalidSessionIdException, StaleElementReferenceException)
     )
-    def batteryinfo(self, params: Union[Dict, List] = None) -> 'Shadowstep':
+    def batteryinfo(self, params: dict[str, Any] | list[Any] | None = None) -> MobileCommands:
         """
         https://github.com/appium/appium-uiautomator2-driver/blob/master/docs/android-mobile-gestures.md#mobile batteryinfo
         Execute mobile: batteryInfo command.
@@ -56,8 +65,8 @@ class MobileCommands:
         Returns:
             Shadowstep: Self for method chaining.
         """
-        self.logger.debug(f"{inspect.currentframe().f_code.co_name}")
-        self._execute('mobile: batteryInfo', params or {})
+        self.logger.debug(f"{get_current_func_name()}")
+        self._execute("mobile: batteryInfo", params)
         return self
 
 
@@ -67,7 +76,7 @@ class MobileCommands:
         raise_exception=ShadowstepException,
         exceptions=(NoSuchDriverException, InvalidSessionIdException, StaleElementReferenceException)
     )
-    def clearelement(self, params: Union[Dict, List] = None) -> 'Shadowstep':
+    def clearelement(self, params: dict[str, Any] | list[Any] | None = None) -> MobileCommands:
         """
         https://github.com/appium/appium-uiautomator2-driver/blob/master/docs/android-mobile-gestures.md#mobile clearelement
         Execute mobile: clearElement command.
@@ -78,8 +87,8 @@ class MobileCommands:
         Returns:
             Shadowstep: Self for method chaining.
         """
-        self.logger.debug(f"{inspect.currentframe().f_code.co_name}")
-        self._execute('mobile: clearElement', params or {})
+        self.logger.debug(f"{get_current_func_name()}")
+        self._execute("mobile: clearElement", params)
         return self
 
 
@@ -89,7 +98,7 @@ class MobileCommands:
         raise_exception=ShadowstepException,
         exceptions=(NoSuchDriverException, InvalidSessionIdException, StaleElementReferenceException)
     )
-    def deviceinfo(self, params: Union[Dict, List] = None) -> 'Shadowstep':
+    def deviceinfo(self, params: dict[str, Any] | list[Any] | None = None) -> MobileCommands:
         """
         https://github.com/appium/appium-uiautomator2-driver/blob/master/docs/android-mobile-gestures.md#mobile deviceinfo
         Execute mobile: deviceInfo command.
@@ -100,8 +109,8 @@ class MobileCommands:
         Returns:
             Shadowstep: Self for method chaining.
         """
-        self.logger.debug(f"{inspect.currentframe().f_code.co_name}")
-        self._execute('mobile: deviceInfo', params or {})
+        self.logger.debug(f"{get_current_func_name()}")
+        self._execute("mobile: deviceInfo", params)
         return self
 
 
@@ -111,7 +120,7 @@ class MobileCommands:
         raise_exception=ShadowstepException,
         exceptions=(NoSuchDriverException, InvalidSessionIdException, StaleElementReferenceException)
     )
-    def fingerprint(self, params: Union[Dict, List] = None) -> 'Shadowstep':
+    def fingerprint(self, params: dict[str, Any] | list[Any] | None = None) -> MobileCommands:
         """
         https://github.com/appium/appium-uiautomator2-driver/blob/master/docs/android-mobile-gestures.md#mobile fingerprint
         Execute mobile: fingerprint command.
@@ -122,8 +131,8 @@ class MobileCommands:
         Returns:
             Shadowstep: Self for method chaining.
         """
-        self.logger.debug(f"{inspect.currentframe().f_code.co_name}")
-        self._execute('mobile: fingerprint', params or {})
+        self.logger.debug(f"{get_current_func_name()}")
+        self._execute("mobile: fingerprint", params)
         return self
 
 
@@ -133,7 +142,7 @@ class MobileCommands:
         raise_exception=ShadowstepException,
         exceptions=(NoSuchDriverException, InvalidSessionIdException, StaleElementReferenceException)
     )
-    def get_clipboard(self, params: Union[Dict, List] = None) -> 'Shadowstep':
+    def get_clipboard(self, params: dict[str, Any] | list[Any] | None = None) -> MobileCommands:
         """
         https://github.com/appium/appium-uiautomator2-driver/blob/master/docs/android-mobile-gestures.md#mobile getclipboard
         Execute mobile: getClipboard command.
@@ -144,8 +153,8 @@ class MobileCommands:
         Returns:
             Shadowstep: Self for method chaining.
         """
-        self.logger.debug(f"{inspect.currentframe().f_code.co_name}")
-        self._execute('mobile: getClipboard', params or {})
+        self.logger.debug(f"{get_current_func_name()}")
+        self._execute("mobile: getClipboard", params)
         return self
 
 
@@ -155,7 +164,7 @@ class MobileCommands:
         raise_exception=ShadowstepException,
         exceptions=(NoSuchDriverException, InvalidSessionIdException, StaleElementReferenceException)
     )
-    def get_current_activity(self, params: Union[Dict, List] = None) -> 'Shadowstep':
+    def get_current_activity(self, params: dict[str, Any] | list[Any] | None = None) -> MobileCommands:
         """
         https://github.com/appium/appium-uiautomator2-driver/blob/master/docs/android-mobile-gestures.md#mobile getcurrentactivity
         Execute mobile: getCurrentActivity command.
@@ -166,8 +175,8 @@ class MobileCommands:
         Returns:
             Shadowstep: Self for method chaining.
         """
-        self.logger.debug(f"{inspect.currentframe().f_code.co_name}")
-        self._execute('mobile: getCurrentActivity', params or {})
+        self.logger.debug(f"{get_current_func_name()}")
+        self._execute("mobile: getCurrentActivity", params)
         return self
 
 
@@ -177,7 +186,7 @@ class MobileCommands:
         raise_exception=ShadowstepException,
         exceptions=(NoSuchDriverException, InvalidSessionIdException, StaleElementReferenceException)
     )
-    def get_currentpackage(self, params: Union[Dict, List] = None) -> 'Shadowstep':
+    def get_currentpackage(self, params: dict[str, Any] | list[Any] | None = None) -> MobileCommands:
         """
         https://github.com/appium/appium-uiautomator2-driver/blob/master/docs/android-mobile-gestures.md#mobile getcurrentpackage
         Execute mobile: getCurrentPackage command.
@@ -188,8 +197,8 @@ class MobileCommands:
         Returns:
             Shadowstep: Self for method chaining.
         """
-        self.logger.debug(f"{inspect.currentframe().f_code.co_name}")
-        self._execute('mobile: getCurrentPackage', params or {})
+        self.logger.debug(f"{get_current_func_name()}")
+        self._execute("mobile: getCurrentPackage", params)
         return self
 
 
@@ -199,7 +208,7 @@ class MobileCommands:
         raise_exception=ShadowstepException,
         exceptions=(NoSuchDriverException, InvalidSessionIdException, StaleElementReferenceException)
     )
-    def get_devicetime(self, params: Union[Dict, List] = None) -> 'Shadowstep':
+    def get_devicetime(self, params: dict[str, Any] | list[Any] | None = None) -> MobileCommands:
         """
         https://github.com/appium/appium-uiautomator2-driver/blob/master/docs/android-mobile-gestures.md#mobile getdevicetime
         Execute mobile: getDeviceTime command.
@@ -210,8 +219,8 @@ class MobileCommands:
         Returns:
             Shadowstep: Self for method chaining.
         """
-        self.logger.debug(f"{inspect.currentframe().f_code.co_name}")
-        self._execute('mobile: getDeviceTime', params or {})
+        self.logger.debug(f"{get_current_func_name()}")
+        self._execute("mobile: getDeviceTime", params)
         return self
 
 
@@ -221,7 +230,7 @@ class MobileCommands:
         raise_exception=ShadowstepException,
         exceptions=(NoSuchDriverException, InvalidSessionIdException, StaleElementReferenceException)
     )
-    def get_performancedata(self, params: Union[Dict, List] = None) -> 'Shadowstep':
+    def get_performancedata(self, params: dict[str, Any] | list[Any] | None = None) -> MobileCommands:
         """
         https://github.com/appium/appium-uiautomator2-driver/blob/master/docs/android-mobile-gestures.md#mobile getperformancedata
         Execute mobile: getPerformanceData command.
@@ -232,8 +241,8 @@ class MobileCommands:
         Returns:
             Shadowstep: Self for method chaining.
         """
-        self.logger.debug(f"{inspect.currentframe().f_code.co_name}")
-        self._execute('mobile: getPerformanceData', params or {})
+        self.logger.debug(f"{get_current_func_name()}")
+        self._execute("mobile: getPerformanceData", params)
         return self
 
 
@@ -243,7 +252,7 @@ class MobileCommands:
         raise_exception=ShadowstepException,
         exceptions=(NoSuchDriverException, InvalidSessionIdException, StaleElementReferenceException)
     )
-    def get_performancedatatypes(self, params: Union[Dict, List] = None) -> 'Shadowstep':
+    def get_performancedatatypes(self, params: dict[str, Any] | list[Any] | None = None) -> MobileCommands:
         """
         https://github.com/appium/appium-uiautomator2-driver/blob/master/docs/android-mobile-gestures.md#mobile getperformancedatatypes
         Execute mobile: getPerformanceDataTypes command.
@@ -254,8 +263,8 @@ class MobileCommands:
         Returns:
             Shadowstep: Self for method chaining.
         """
-        self.logger.debug(f"{inspect.currentframe().f_code.co_name}")
-        self._execute('mobile: getPerformanceDataTypes', params or {})
+        self.logger.debug(f"{get_current_func_name()}")
+        self._execute("mobile: getPerformanceDataTypes", params)
         return self
 
 
@@ -265,7 +274,7 @@ class MobileCommands:
         raise_exception=ShadowstepException,
         exceptions=(NoSuchDriverException, InvalidSessionIdException, StaleElementReferenceException)
     )
-    def get_settings(self, params: Union[Dict, List] = None) -> 'Shadowstep':
+    def get_settings(self, params: dict[str, Any] | list[Any] | None = None) -> MobileCommands:
         """
         https://github.com/appium/appium-uiautomator2-driver/blob/master/docs/android-mobile-gestures.md#mobile getsettings
         Execute mobile: getSettings command.
@@ -276,8 +285,8 @@ class MobileCommands:
         Returns:
             Shadowstep: Self for method chaining.
         """
-        self.logger.debug(f"{inspect.currentframe().f_code.co_name}")
-        self._execute('mobile: getSettings', params or {})
+        self.logger.debug(f"{get_current_func_name()}")
+        self._execute("mobile: getSettings", params)
         return self
 
 
@@ -287,7 +296,7 @@ class MobileCommands:
         raise_exception=ShadowstepException,
         exceptions=(NoSuchDriverException, InvalidSessionIdException, StaleElementReferenceException)
     )
-    def hide_keyboard(self, params: Union[Dict, List] = None) -> 'Shadowstep':
+    def hide_keyboard(self, params: dict[str, Any] | list[Any] | None = None) -> MobileCommands:
         """
         https://github.com/appium/appium-uiautomator2-driver/blob/master/docs/android-mobile-gestures.md#mobile hidekeyboard
         Execute mobile: hideKeyboard command.
@@ -298,8 +307,8 @@ class MobileCommands:
         Returns:
             Shadowstep: Self for method chaining.
         """
-        self.logger.debug(f"{inspect.currentframe().f_code.co_name}")
-        self._execute('mobile: hideKeyboard', params or {})
+        self.logger.debug(f"{get_current_func_name()}")
+        self._execute("mobile: hideKeyboard", params)
         return self
 
 
@@ -309,7 +318,7 @@ class MobileCommands:
         raise_exception=ShadowstepException,
         exceptions=(NoSuchDriverException, InvalidSessionIdException, StaleElementReferenceException)
     )
-    def install_app(self, params: Union[Dict, List] = None) -> 'Shadowstep':
+    def install_app(self, params: dict[str, Any] | list[Any] | None = None) -> MobileCommands:
         """
         https://github.com/appium/appium-uiautomator2-driver/blob/master/docs/android-mobile-gestures.md#mobile installapp
         Execute mobile: installApp command.
@@ -320,8 +329,8 @@ class MobileCommands:
         Returns:
             Shadowstep: Self for method chaining.
         """
-        self.logger.debug(f"{inspect.currentframe().f_code.co_name}")
-        self._execute('mobile: installApp', params or {})
+        self.logger.debug(f"{get_current_func_name()}")
+        self._execute("mobile: installApp", params)
         return self
 
 
@@ -331,7 +340,7 @@ class MobileCommands:
         raise_exception=ShadowstepException,
         exceptions=(NoSuchDriverException, InvalidSessionIdException, StaleElementReferenceException)
     )
-    def is_appinstalled(self, params: Union[Dict, List] = None) -> 'Shadowstep':
+    def is_appinstalled(self, params: dict[str, Any] | list[Any] | None = None) -> MobileCommands:
         """
         https://github.com/appium/appium-uiautomator2-driver/blob/master/docs/android-mobile-gestures.md#mobile isappinstalled
         Execute mobile: isAppInstalled command.
@@ -342,8 +351,8 @@ class MobileCommands:
         Returns:
             Shadowstep: Self for method chaining.
         """
-        self.logger.debug(f"{inspect.currentframe().f_code.co_name}")
-        self._execute('mobile: isAppInstalled', params or {})
+        self.logger.debug(f"{get_current_func_name()}")
+        self._execute("mobile: isAppInstalled", params)
         return self
 
 
@@ -353,7 +362,7 @@ class MobileCommands:
         raise_exception=ShadowstepException,
         exceptions=(NoSuchDriverException, InvalidSessionIdException, StaleElementReferenceException)
     )
-    def longpress_key(self, params: Union[Dict, List] = None) -> 'Shadowstep':
+    def longpress_key(self, params: dict[str, Any] | list[Any] | None = None) -> MobileCommands:
         """
         https://github.com/appium/appium-uiautomator2-driver/blob/master/docs/android-mobile-gestures.md#mobile longpresskey
         Execute mobile: longPressKey command.
@@ -364,8 +373,8 @@ class MobileCommands:
         Returns:
             Shadowstep: Self for method chaining.
         """
-        self.logger.debug(f"{inspect.currentframe().f_code.co_name}")
-        self._execute('mobile: longPressKey', params or {})
+        self.logger.debug(f"{get_current_func_name()}")
+        self._execute("mobile: longPressKey", params)
         return self
 
 
@@ -375,7 +384,7 @@ class MobileCommands:
         raise_exception=ShadowstepException,
         exceptions=(NoSuchDriverException, InvalidSessionIdException, StaleElementReferenceException)
     )
-    def open_notifications(self, params: Union[Dict, List] = None) -> 'Shadowstep':
+    def open_notifications(self, params: dict[str, Any] | list[Any] | None = None) -> MobileCommands:
         """
         https://github.com/appium/appium-uiautomator2-driver/blob/master/docs/android-mobile-gestures.md#mobile opennotifications
         Execute mobile: openNotifications command.
@@ -386,8 +395,8 @@ class MobileCommands:
         Returns:
             Shadowstep: Self for method chaining.
         """
-        self.logger.debug(f"{inspect.currentframe().f_code.co_name}")
-        self._execute('mobile: openNotifications', params or {})
+        self.logger.debug(f"{get_current_func_name()}")
+        self._execute("mobile: openNotifications", params)
         return self
 
 
@@ -397,7 +406,7 @@ class MobileCommands:
         raise_exception=ShadowstepException,
         exceptions=(NoSuchDriverException, InvalidSessionIdException, StaleElementReferenceException)
     )
-    def open_settings(self, params: Union[Dict, List] = None) -> 'Shadowstep':
+    def open_settings(self, params: dict[str, Any] | list[Any] | None = None) -> MobileCommands:
         """
         https://github.com/appium/appium-uiautomator2-driver/blob/master/docs/android-mobile-gestures.md#mobile opensettings
         Execute mobile: openSettings command.
@@ -408,8 +417,8 @@ class MobileCommands:
         Returns:
             Shadowstep: Self for method chaining.
         """
-        self.logger.debug(f"{inspect.currentframe().f_code.co_name}")
-        self._execute('mobile: openSettings', params or {})
+        self.logger.debug(f"{get_current_func_name()}")
+        self._execute("mobile: openSettings", params)
         return self
 
 
@@ -419,7 +428,7 @@ class MobileCommands:
         raise_exception=ShadowstepException,
         exceptions=(NoSuchDriverException, InvalidSessionIdException, StaleElementReferenceException)
     )
-    def press_key(self, params: Union[Dict, List] = None) -> 'Shadowstep':
+    def press_key(self, params: dict[str, Any] | list[Any] | None = None) -> MobileCommands:
         """
         https://github.com/appium/appium-uiautomator2-driver/blob/master/docs/android-mobile-gestures.md#mobile presskey
         Execute mobile: pressKey command.
@@ -430,8 +439,8 @@ class MobileCommands:
         Returns:
             Shadowstep: Self for method chaining.
         """
-        self.logger.debug(f"{inspect.currentframe().f_code.co_name}")
-        self._execute('mobile: pressKey', params or {})
+        self.logger.debug(f"{get_current_func_name()}")
+        self._execute("mobile: pressKey", params)
         return self
 
 
@@ -441,7 +450,7 @@ class MobileCommands:
         raise_exception=ShadowstepException,
         exceptions=(NoSuchDriverException, InvalidSessionIdException, StaleElementReferenceException)
     )
-    def query_appstate(self, params: Union[Dict, List] = None) -> 'Shadowstep':
+    def query_appstate(self, params: dict[str, Any] | list[Any] | None = None) -> MobileCommands:
         """
         https://github.com/appium/appium-uiautomator2-driver/blob/master/docs/android-mobile-gestures.md#mobile queryappstate
         Execute mobile: queryAppState command.
@@ -452,8 +461,8 @@ class MobileCommands:
         Returns:
             Shadowstep: Self for method chaining.
         """
-        self.logger.debug(f"{inspect.currentframe().f_code.co_name}")
-        self._execute('mobile: queryAppState', params or {})
+        self.logger.debug(f"{get_current_func_name()}")
+        self._execute("mobile: queryAppState", params)
         return self
 
 
@@ -463,7 +472,7 @@ class MobileCommands:
         raise_exception=ShadowstepException,
         exceptions=(NoSuchDriverException, InvalidSessionIdException, StaleElementReferenceException)
     )
-    def remove_app(self, params: Union[Dict, List] = None) -> 'Shadowstep':
+    def remove_app(self, params: dict[str, Any] | list[Any] | None = None) -> MobileCommands:
         """
         https://github.com/appium/appium-uiautomator2-driver/blob/master/docs/android-mobile-gestures.md#mobile removeapp
         Execute mobile: removeApp command.
@@ -474,8 +483,8 @@ class MobileCommands:
         Returns:
             Shadowstep: Self for method chaining.
         """
-        self.logger.debug(f"{inspect.currentframe().f_code.co_name}")
-        self._execute('mobile: removeApp', params or {})
+        self.logger.debug(f"{get_current_func_name()}")
+        self._execute("mobile: removeApp", params)
         return self
 
 
@@ -485,7 +494,7 @@ class MobileCommands:
         raise_exception=ShadowstepException,
         exceptions=(NoSuchDriverException, InvalidSessionIdException, StaleElementReferenceException)
     )
-    def replaceelementvalue(self, params: Union[Dict, List] = None) -> 'Shadowstep':
+    def replaceelementvalue(self, params: dict[str, Any] | list[Any] | None = None) -> MobileCommands:
         """
         https://github.com/appium/appium-uiautomator2-driver/blob/master/docs/android-mobile-gestures.md#mobile replaceelementvalue
         Execute mobile: replaceElementValue command.
@@ -496,8 +505,8 @@ class MobileCommands:
         Returns:
             Shadowstep: Self for method chaining.
         """
-        self.logger.debug(f"{inspect.currentframe().f_code.co_name}")
-        self._execute('mobile: replaceElementValue', params or {})
+        self.logger.debug(f"{get_current_func_name()}")
+        self._execute("mobile: replaceElementValue", params)
         return self
 
 
@@ -507,7 +516,7 @@ class MobileCommands:
         raise_exception=ShadowstepException,
         exceptions=(NoSuchDriverException, InvalidSessionIdException, StaleElementReferenceException)
     )
-    def scroll_back_to(self, params: Union[Dict, List] = None) -> 'Shadowstep':
+    def scroll_back_to(self, params: dict[str, Any] | list[Any] | None = None) -> MobileCommands:
         """
         https://github.com/appium/appium-uiautomator2-driver/blob/master/docs/android-mobile-gestures.md#mobile scrollbackto
         Execute mobile: scrollBackTo command.
@@ -518,8 +527,8 @@ class MobileCommands:
         Returns:
             Shadowstep: Self for method chaining.
         """
-        self.logger.debug(f"{inspect.currentframe().f_code.co_name}")
-        self._execute('mobile: scrollBackTo', params or {})
+        self.logger.debug(f"{get_current_func_name()}")
+        self._execute("mobile: scrollBackTo", params)
         return self
 
 
@@ -529,7 +538,7 @@ class MobileCommands:
         raise_exception=ShadowstepException,
         exceptions=(NoSuchDriverException, InvalidSessionIdException, StaleElementReferenceException)
     )
-    def send_sms(self, params: Union[Dict, List] = None) -> 'Shadowstep':
+    def send_sms(self, params: dict[str, Any] | list[Any] | None = None) -> MobileCommands:
         """
         https://github.com/appium/appium-uiautomator2-driver/blob/master/docs/android-mobile-gestures.md#mobile sendsms
         Execute mobile: sendSMS command.
@@ -540,8 +549,8 @@ class MobileCommands:
         Returns:
             Shadowstep: Self for method chaining.
         """
-        self.logger.debug(f"{inspect.currentframe().f_code.co_name}")
-        self._execute('mobile: sendSMS', params or {})
+        self.logger.debug(f"{get_current_func_name()}")
+        self._execute("mobile: sendSMS", params)
         return self
 
 
@@ -551,7 +560,7 @@ class MobileCommands:
         raise_exception=ShadowstepException,
         exceptions=(NoSuchDriverException, InvalidSessionIdException, StaleElementReferenceException)
     )
-    def set_clipboard(self, params: Union[Dict, List] = None) -> 'Shadowstep':
+    def set_clipboard(self, params: dict[str, Any] | list[Any] | None = None) -> MobileCommands:
         """
         https://github.com/appium/appium-uiautomator2-driver/blob/master/docs/android-mobile-gestures.md#mobile setclipboard
         Execute mobile: setClipboard command.
@@ -562,8 +571,8 @@ class MobileCommands:
         Returns:
             Shadowstep: Self for method chaining.
         """
-        self.logger.debug(f"{inspect.currentframe().f_code.co_name}")
-        self._execute('mobile: setClipboard', params or {})
+        self.logger.debug(f"{get_current_func_name()}")
+        self._execute("mobile: setClipboard", params)
         return self
 
 
@@ -573,7 +582,7 @@ class MobileCommands:
         raise_exception=ShadowstepException,
         exceptions=(NoSuchDriverException, InvalidSessionIdException, StaleElementReferenceException)
     )
-    def set_text(self, params: Union[Dict, List] = None) -> 'Shadowstep':
+    def set_text(self, params: dict[str, Any] | list[Any] | None = None) -> MobileCommands:
         """
         https://github.com/appium/appium-uiautomator2-driver/blob/master/docs/android-mobile-gestures.md#mobile settext
         Execute mobile: setText command.
@@ -584,8 +593,8 @@ class MobileCommands:
         Returns:
             Shadowstep: Self for method chaining.
         """
-        self.logger.debug(f"{inspect.currentframe().f_code.co_name}")
-        self._execute('mobile: setText', params or {})
+        self.logger.debug(f"{get_current_func_name()}")
+        self._execute("mobile: setText", params)
         return self
 
 
@@ -595,7 +604,7 @@ class MobileCommands:
         raise_exception=ShadowstepException,
         exceptions=(NoSuchDriverException, InvalidSessionIdException, StaleElementReferenceException)
     )
-    def shell(self, params: Union[Dict, List] = None) -> 'Shadowstep':
+    def shell(self, params: dict[str, Any] | list[Any] | None = None) -> MobileCommands:
         """
         https://github.com/appium/appium-uiautomator2-driver/blob/master/docs/android-mobile-gestures.md#mobile shell
         Execute mobile: shell command.
@@ -606,8 +615,8 @@ class MobileCommands:
         Returns:
             Shadowstep: Self for method chaining.
         """
-        self.logger.debug(f"{inspect.currentframe().f_code.co_name}")
-        self._execute('mobile: shell', params or {})
+        self.logger.debug(f"{get_current_func_name()}")
+        self._execute("mobile: shell", params)
         return self
 
 
@@ -617,7 +626,7 @@ class MobileCommands:
         raise_exception=ShadowstepException,
         exceptions=(NoSuchDriverException, InvalidSessionIdException, StaleElementReferenceException)
     )
-    def start_activity(self, params: Union[Dict, List] = None) -> 'Shadowstep':
+    def start_activity(self, params: dict[str, Any] | list[Any] | None = None) -> MobileCommands:
         """
         https://github.com/appium/appium-uiautomator2-driver/blob/master/docs/android-mobile-gestures.md#mobile startactivity
         Execute mobile: startActivity command.
@@ -628,8 +637,8 @@ class MobileCommands:
         Returns:
             Shadowstep: Self for method chaining.
         """
-        self.logger.debug(f"{inspect.currentframe().f_code.co_name}")
-        self._execute('mobile: startActivity', params or {})
+        self.logger.debug(f"{get_current_func_name()}")
+        self._execute("mobile: startActivity", params)
         return self
 
 
@@ -639,7 +648,7 @@ class MobileCommands:
         raise_exception=ShadowstepException,
         exceptions=(NoSuchDriverException, InvalidSessionIdException, StaleElementReferenceException)
     )
-    def start_logsbroadcast(self, params: Union[Dict, List] = None) -> 'Shadowstep':
+    def start_logsbroadcast(self, params: dict[str, Any] | list[Any] | None = None) -> MobileCommands:
         """
         https://github.com/appium/appium-uiautomator2-driver/blob/master/docs/android-mobile-gestures.md#mobile startlogsbroadcast
         Execute mobile: startLogsBroadcast command.
@@ -650,8 +659,8 @@ class MobileCommands:
         Returns:
             Shadowstep: Self for method chaining.
         """
-        self.logger.debug(f"{inspect.currentframe().f_code.co_name}")
-        self._execute('mobile: startLogsBroadcast', params or {})
+        self.logger.debug(f"{get_current_func_name()}")
+        self._execute("mobile: startLogsBroadcast", params)
         return self
 
 
@@ -661,7 +670,7 @@ class MobileCommands:
         raise_exception=ShadowstepException,
         exceptions=(NoSuchDriverException, InvalidSessionIdException, StaleElementReferenceException)
     )
-    def stop_logsbroadcast(self, params: Union[Dict, List] = None) -> 'Shadowstep':
+    def stop_logsbroadcast(self, params: dict[str, Any] | list[Any] | None = None) -> MobileCommands:
         """
         https://github.com/appium/appium-uiautomator2-driver/blob/master/docs/android-mobile-gestures.md#mobile stoplogsbroadcast
         Execute mobile: stopLogsBroadcast command.
@@ -672,8 +681,8 @@ class MobileCommands:
         Returns:
             Shadowstep: Self for method chaining.
         """
-        self.logger.debug(f"{inspect.currentframe().f_code.co_name}")
-        self._execute('mobile: stopLogsBroadcast', params or {})
+        self.logger.debug(f"{get_current_func_name()}")
+        self._execute("mobile: stopLogsBroadcast", params)
         return self
 
 
@@ -683,7 +692,7 @@ class MobileCommands:
         raise_exception=ShadowstepException,
         exceptions=(NoSuchDriverException, InvalidSessionIdException, StaleElementReferenceException)
     )
-    def terminate_app(self, params: Union[Dict, List] = None) -> 'Shadowstep':
+    def terminate_app(self, params: dict[str, Any] | list[Any] | None = None) -> MobileCommands:
         """
         https://github.com/appium/appium-uiautomator2-driver/blob/master/docs/android-mobile-gestures.md#mobile terminateapp
         Execute mobile: terminateApp command.
@@ -694,8 +703,8 @@ class MobileCommands:
         Returns:
             Shadowstep: Self for method chaining.
         """
-        self.logger.debug(f"{inspect.currentframe().f_code.co_name}")
-        self._execute('mobile: terminateApp', params or {})
+        self.logger.debug(f"{get_current_func_name()}")
+        self._execute("mobile: terminateApp", params)
         return self
 
 
@@ -705,7 +714,7 @@ class MobileCommands:
         raise_exception=ShadowstepException,
         exceptions=(NoSuchDriverException, InvalidSessionIdException, StaleElementReferenceException)
     )
-    def togglelocationservices(self, params: Union[Dict, List] = None) -> 'Shadowstep':
+    def togglelocationservices(self, params: dict[str, Any] | list[Any] | None = None) -> MobileCommands:
         """
         https://github.com/appium/appium-uiautomator2-driver/blob/master/docs/android-mobile-gestures.md#mobile togglelocationservices
         Execute mobile: toggleLocationServices command.
@@ -716,8 +725,8 @@ class MobileCommands:
         Returns:
             Shadowstep: Self for method chaining.
         """
-        self.logger.debug(f"{inspect.currentframe().f_code.co_name}")
-        self._execute('mobile: toggleLocationServices', params or {})
+        self.logger.debug(f"{get_current_func_name()}")
+        self._execute("mobile: toggleLocationServices", params)
         return self
 
 
@@ -727,7 +736,7 @@ class MobileCommands:
         raise_exception=ShadowstepException,
         exceptions=(NoSuchDriverException, InvalidSessionIdException, StaleElementReferenceException)
     )
-    def update_settings(self, params: Union[Dict, List] = None) -> 'Shadowstep':
+    def update_settings(self, params: dict[str, Any] | list[Any] | None = None) -> MobileCommands:
         """
         https://github.com/appium/appium-uiautomator2-driver/blob/master/docs/android-mobile-gestures.md#mobile updatesettings
         Execute mobile: updateSettings command.
@@ -738,8 +747,8 @@ class MobileCommands:
         Returns:
             Shadowstep: Self for method chaining.
         """
-        self.logger.debug(f"{inspect.currentframe().f_code.co_name}")
-        self._execute('mobile: updateSettings', params or {})
+        self.logger.debug(f"{get_current_func_name()}")
+        self._execute("mobile: updateSettings", params)
         return self
 
 
@@ -749,7 +758,7 @@ class MobileCommands:
         raise_exception=ShadowstepException,
         exceptions=(NoSuchDriverException, InvalidSessionIdException, StaleElementReferenceException)
     )
-    def get_text(self, params: Union[Dict, List] = None) -> 'Shadowstep':
+    def get_text(self, params: dict[str, Any] | list[Any] | None = None) -> MobileCommands:
         """
         https://github.com/appium/appium-uiautomator2-driver/blob/master/docs/android-mobile-gestures.md#mobile gettext
         Execute mobile: getText command.
@@ -760,8 +769,8 @@ class MobileCommands:
         Returns:
             Shadowstep: Self for method chaining.
         """
-        self.logger.debug(f"{inspect.currentframe().f_code.co_name}")
-        self._execute('mobile: getText', params or {})
+        self.logger.debug(f"{get_current_func_name()}")
+        self._execute("mobile: getText", params)
         return self
 
 
@@ -771,7 +780,7 @@ class MobileCommands:
         raise_exception=ShadowstepException,
         exceptions=(NoSuchDriverException, InvalidSessionIdException, StaleElementReferenceException)
     )
-    def performeditoraction(self, params: Union[Dict, List] = None) -> 'Shadowstep':
+    def performeditoraction(self, params: dict[str, Any] | list[Any] | None = None) -> MobileCommands:
         """
         https://github.com/appium/appium-uiautomator2-driver/blob/master/docs/android-mobile-gestures.md#mobile performeditoraction
         Execute mobile: performEditorAction command.
@@ -782,11 +791,13 @@ class MobileCommands:
         Returns:
             Shadowstep: Self for method chaining.
         """
-        self.logger.debug(f"{inspect.currentframe().f_code.co_name}")
-        self._execute('mobile: performEditorAction', params or {})
+        self.logger.debug(f"{get_current_func_name()}")
+        self._execute("mobile: performEditorAction", params)
         return self
 
-    def _execute(self, name: str, params: Union[dict, list]) -> None:
+    def _execute(self, name: str, params: dict[str, Any] | list[Any] | None) -> None:
         # https://github.com/appium/appium-uiautomator2-driver/blob/master/docs/android-mobile-gestures.md
         self.driver = WebDriverSingleton.get_driver()
-        self.driver.execute_script(name, params)
+        if self.driver is None:
+            raise ShadowstepException("WebDriver is not available")
+        self.driver.execute_script(name, params or {})
