@@ -1,19 +1,23 @@
 # shadowstep/terminal/adb.py
+from __future__ import annotations
 
-import inspect
-
-from appium.webdriver.webdriver import WebDriver
 import logging
-logger = logging.getLogger(__name__)
 import os
 import re
 import subprocess
 import sys
 import time
 import traceback
-from typing import Dict, Union, Tuple, Optional, Any, List
+from typing import TYPE_CHECKING
 
-from shadowstep.utils import operations
+from appium.webdriver.webdriver import WebDriver
+
+from shadowstep.utils.utils import get_current_func_name, grep_pattern
+
+if TYPE_CHECKING:
+    from shadowstep.base import ShadowstepBase
+
+logger = logging.getLogger(__name__)
 
 
 class Adb:
@@ -22,16 +26,15 @@ class Adb:
     Use only if Appium server is running locally where the test is being performed
     """
 
-    base = None
-    driver = None
+    base: ShadowstepBase
+    driver: WebDriver
 
-    def __init__(self, base: "Shadowstep"):
-        self.base: "Shadowstep" = base
+    def __init__(self, base: ShadowstepBase):
+        self.base: ShadowstepBase = base
         self.driver: WebDriver = base.driver
 
-        
     @staticmethod
-    def get_devices() -> Union[List[str], None]:
+    def get_devices() -> list[str]:
         """
         Retrieve a list of connected devices via ADB.
 
@@ -39,34 +42,34 @@ class Adb:
             Union[List[str], None]
                 A list of connected device identifiers (UUIDs) or None if no devices are found or an error occurs.
         """
-        logger.info(f"{inspect.currentframe().f_code.co_name}")
+        logger.info(f"{get_current_func_name()}")
 
         # Определение команды для выполнения с помощью adb для получения списка устройств
-        command = ['adb', 'devices']
+        command = ["adb", "devices"]
 
         try:
             # Выполнение команды и получение вывода
-            response = str(subprocess.check_output(command))
+            response = str(subprocess.check_output(command))  # noqa: S603
 
             # Извлечение списка устройств из полученного вывода с использованием регулярных выражений
-            devices_list = re.findall(r'(\d+\.\d+\.\d+\.\d+:\d+|\d+)', response)
+            devices_list = re.findall(r"(\d+\.\d+\.\d+\.\d+:\d+|\d+)", response)
 
             try:
                 # Возвращение первого устройства из списка (UUID подключенного устройства Android)
-                logger.info(f"{inspect.currentframe().f_code.co_name} > {devices_list}")
+                logger.info(f"{get_current_func_name()} > {devices_list}")
                 return devices_list
             except IndexError:
-                logger.error(f"{inspect.currentframe().f_code.co_name} > None")
+                logger.error(f"{get_current_func_name()} > None")
                 logger.error("No connected devices")
-                return None
-        except subprocess.CalledProcessError as e:
-            logger.error(f"{inspect.currentframe().f_code.co_name} > None")
+                return []
+        except subprocess.CalledProcessError:
+            logger.error(f"{get_current_func_name()} > None")
             traceback_info = "".join(traceback.format_tb(sys.exc_info()[2]))
             logger.error(f"{sys.exc_info()[0]}\n{traceback_info}{sys.exc_info()[1]}")
-            return None
+            return []
 
     @staticmethod
-    def get_device_model(udid: str = None) -> Union[str, None]:
+    def get_device_model(udid: str) -> str:
         """
         Retrieve the model of the connected device using ADB.
 
@@ -79,24 +82,23 @@ class Adb:
             Union[str, None]
                 The model of the device as a string, or None if an error occurs or the model cannot be retrieved.
         """
-        logger.info(f"{inspect.currentframe().f_code.co_name} < {udid}")
-        s_udid = f"-s {udid}" if udid else ""
-        command = [f"adb {s_udid}", "shell", "getprop", "ro.product.model"]
+        logger.info(f"{get_current_func_name()} < {udid}")
+        command = ["adb", "-s", f"{udid}", "shell", "getprop", "ro.product.model"] if udid else ["adb", "shell", "getprop", "ro.product.model"]
         try:
             # Выполнение команды и получение вывода
-            model = subprocess.check_output(command)
+            model = subprocess.check_output(command)  # noqa: S603
             # Преобразование байтовой строки в обычную строку и удаление пробельных символов и символов перевода строки
             model = model.decode().strip()
-            logger.info(f"{inspect.currentframe().f_code.co_name} > {model}")
+            logger.info(f"{get_current_func_name()} > {model}")
             return model
-        except subprocess.CalledProcessError as e:
-            logger.error(f"{inspect.currentframe().f_code.co_name} > None")
+        except subprocess.CalledProcessError:
+            logger.error(f"{get_current_func_name()} > None")
             traceback_info = "".join(traceback.format_tb(sys.exc_info()[2]))
             logger.error(f"{sys.exc_info()[0]}\n{traceback_info}{sys.exc_info()[1]}")
-            return None
+            return ""
 
     @staticmethod
-    def push(source: str, destination: str, udid: str = None) -> bool:
+    def push(source: str, destination: str, udid: str) -> bool:
         """
         Push a file from the local machine to the connected device using ADB.
 
@@ -112,25 +114,24 @@ class Adb:
             bool
                 True if the file was successfully pushed, False otherwise.
         """
-        logger.info(f"{inspect.currentframe().f_code.co_name} < {source=}, {destination=}")
+        logger.info(f"{get_current_func_name()} < {source=}, {destination=}")
 
         if not os.path.exists(source):
             logger.error(f"Source path does not exist: {source=}")
             return False
-        s_udid = f"-s {udid}" if udid else ""
-        command = f"adb {s_udid} push {source} {destination}"
+        command = ["adb", "-s", f"{udid}", "push", f"{source}", f"{destination}"] if udid else ["adb", "push", f"{source}", f"{destination}"]
         try:
-            subprocess.run(command, check=True)
-            logger.info(f"{inspect.currentframe().f_code.co_name} > True")
+            subprocess.run(command, check=True)  # noqa: S603
+            logger.info(f"{get_current_func_name()} > True")
             return True
-        except subprocess.CalledProcessError as e:
-            logger.error(f"{inspect.currentframe().f_code.co_name} > None")
+        except subprocess.CalledProcessError:
+            logger.error(f"{get_current_func_name()} > None")
             traceback_info = "".join(traceback.format_tb(sys.exc_info()[2]))
             logger.error(f"{sys.exc_info()[0]}\n{traceback_info}{sys.exc_info()[1]}")
             return False
 
     @staticmethod
-    def pull(source: str, destination: str, udid: str = None) -> bool:
+    def pull(source: str, destination: str, udid: str) -> bool:
         """
         Pull a file from the connected device to the local machine using ADB.
 
@@ -146,15 +147,14 @@ class Adb:
             bool
                 True if the file was successfully pulled, False otherwise.
         """
-        logger.info(f"{inspect.currentframe().f_code.co_name} < {source=}, {destination=}")
-        s_udid = f"-s {udid}" if udid else ""
-        command = f"adb {s_udid} pull {source} {destination}"
+        logger.info(f"{get_current_func_name()} < {source=}, {destination=}")
+        command = ["adb", "-s", f"{udid}", "pull", f"{source}", f"{destination}"] if udid else ["adb", "pull", f"{source}", f"{destination}"]
         try:
-            subprocess.run(command, check=True)
-            logger.info(f"{inspect.currentframe().f_code.co_name} > True")
+            subprocess.run(command, check=True)  # noqa: S603
+            logger.info(f"{get_current_func_name()} > True")
             return True
-        except subprocess.CalledProcessError as e:
-            logger.error(f"{inspect.currentframe().f_code.co_name} > None")
+        except subprocess.CalledProcessError:
+            logger.error(f"{get_current_func_name()} > None")
             traceback_info = "".join(traceback.format_tb(sys.exc_info()[2]))
             logger.error(f"{sys.exc_info()[0]}\n{traceback_info}{sys.exc_info()[1]}")
             return False
@@ -175,20 +175,19 @@ class Adb:
                 True if the application was successfully installed, False otherwise.
         """
         logger.info(f"install() < {source=}")
-        s_udid = f"-s {udid}" if udid else ""
-        command = f"adb {s_udid} install -r {source}"
+        command = ["adb", "-s", f"{udid}", "install", "-r", f"{source}"] if udid else ["adb", "install", f"{source}"]
         try:
-            subprocess.run(command, check=True)
+            subprocess.run(command, check=True)  # noqa: S603
             logger.info("install() > True")
             return True
-        except subprocess.CalledProcessError as e:
-            logger.error(f"{inspect.currentframe().f_code.co_name} > None")
+        except subprocess.CalledProcessError:
+            logger.error(f"{get_current_func_name()} > None")
             traceback_info = "".join(traceback.format_tb(sys.exc_info()[2]))
             logger.error(f"{sys.exc_info()[0]}\n{traceback_info}{sys.exc_info()[1]}")
             return False
 
     @staticmethod
-    def is_app_installed(package) -> bool:
+    def is_app_installed(package: str) -> bool:
         """
         Check if the specified package is installed on the connected device.
 
@@ -204,15 +203,15 @@ class Adb:
 
         command = "adb shell pm list packages"
         try:
-            result = subprocess.check_output(command, shell=True).decode().strip()
+            result = subprocess.check_output(command, shell=True).decode().strip()  # noqa: S602
             # Фильтруем пакеты
-            if any([line.strip().endswith(package) for line in result.splitlines()]):
+            if any(line.strip().endswith(package) for line in result.splitlines()):
                 logger.info("install() > True")
                 return True
             logger.info("install() > False")
             return False
-        except subprocess.CalledProcessError as e:
-            logger.error(f"{inspect.currentframe().f_code.co_name} > None")
+        except subprocess.CalledProcessError:
+            logger.error(f"{get_current_func_name()} > None")
             traceback_info = "".join(traceback.format_tb(sys.exc_info()[2]))
             logger.error(f"{sys.exc_info()[0]}\n{traceback_info}{sys.exc_info()[1]}")
             return False
@@ -232,13 +231,13 @@ class Adb:
         """
         logger.info(f"uninstall_app() < {package=}")
 
-        command = ['adb', 'uninstall', package]
+        command = ["adb", "uninstall", package]
         try:
-            subprocess.run(command, check=True)
+            subprocess.run(command, check=True)  # noqa: S603
             logger.info("uninstall_app() > True")
             return True
-        except subprocess.CalledProcessError as e:
-            logger.error(f"{inspect.currentframe().f_code.co_name} > None")
+        except subprocess.CalledProcessError:
+            logger.error(f"{get_current_func_name()} > None")
             traceback_info = "".join(traceback.format_tb(sys.exc_info()[2]))
             logger.error(f"{sys.exc_info()[0]}\n{traceback_info}{sys.exc_info()[1]}")
             return False
@@ -260,19 +259,19 @@ class Adb:
         """
         logger.info(f"start_activity() < {package=}, {activity=}")
 
-        command = ['adb', 'shell', 'am', 'start', '-n', f'{package}/{activity}']
+        command = ["adb", "shell", "am", "start", "-n", f"{package}/{activity}"]
         try:
-            subprocess.check_output(command)
+            subprocess.check_output(command)  # noqa: S603
             logger.info("start_activity() > True")
             return True
-        except subprocess.CalledProcessError as e:
-            logger.error(f"{inspect.currentframe().f_code.co_name} > None")
+        except subprocess.CalledProcessError:
+            logger.error(f"{get_current_func_name()} > None")
             traceback_info = "".join(traceback.format_tb(sys.exc_info()[2]))
             logger.error(f"{sys.exc_info()[0]}\n{traceback_info}{sys.exc_info()[1]}")
             return False
 
     @staticmethod
-    def get_current_activity() -> Union[str, None]:
+    def get_current_activity() -> str:
         """
         Retrieve the name of the current activity running on the device.
 
@@ -284,23 +283,23 @@ class Adb:
         logger.info("get_current_activity()")
 
         # Команда для ADB для получения информации о текущих окнах
-        command = ['adb', 'shell', 'dumpsys', 'window', 'windows']
+        command = ["adb", "shell", "dumpsys", "window", "windows"]
 
         try:
             # Выполнение команды и декодирование результата
-            result = subprocess.check_output(command, shell=True).decode().strip()
+            result = subprocess.check_output(command, shell=True).decode().strip()  # noqa: S602
 
             # Определение паттерна для поиска нужной информации в результатах
-            pattern = r'mCurrentFocus|mFocusedApp'
+            pattern = r"mCurrentFocus|mFocusedApp"
 
             # Вызов функции grep_pattern для поиска соответствия паттерну
-            matched_lines = operations.grep_pattern(input_string=result, pattern=pattern)
+            matched_lines = grep_pattern(input_string=result, pattern=pattern)
 
             # Если были найдены соответствующие строки
             if matched_lines:
                 for line in matched_lines:
                     # Поиск имени активити в строке
-                    match = re.search(r'\/([^\/}]*)', line)
+                    match = re.search(r"\/([^\/}]*)", line)
                     if match:
                         # Возвращаем найденное значение, исключая '/'
                         activity_name = match.group(1)
@@ -309,15 +308,15 @@ class Adb:
 
             # Если не удалось найти активити, возвращаем None
             logger.error("get_current_activity() > None")
-            return None
-        except subprocess.CalledProcessError as e:
-            logger.error(f"{inspect.currentframe().f_code.co_name} > None")
+            return ""
+        except subprocess.CalledProcessError:
+            logger.error(f"{get_current_func_name()} > None")
             traceback_info = "".join(traceback.format_tb(sys.exc_info()[2]))
             logger.error(f"{sys.exc_info()[0]}\n{traceback_info}{sys.exc_info()[1]}")
-            return None
+            return ""
 
     @staticmethod
-    def get_current_package() -> Union[str, None]:
+    def get_current_package() -> str:
         """
         Retrieve the name of the current application package running on the device.
 
@@ -329,23 +328,23 @@ class Adb:
         logger.info("get_current_app_package()")
 
         # Команда для ADB для получения информации о текущих окнах
-        command = ['adb', 'shell', 'dumpsys', 'window', 'windows']
+        command = ["adb", "shell", "dumpsys", "window", "windows"]
 
         try:
             # Выполнение команды и декодирование результата
-            result = subprocess.check_output(command, shell=True).decode().strip()
+            result = subprocess.check_output(command, shell=True).decode().strip()  # noqa: S602
 
             # Определение паттерна для поиска нужной информации в результатах
-            pattern = r'mCurrentFocus|mFocusedApp'
+            pattern = r"mCurrentFocus|mFocusedApp"
 
             # Вызов функции grep_pattern для поиска соответствия паттерну
-            matched_lines = operations.grep_pattern(input_string=result, pattern=pattern)
+            matched_lines = grep_pattern(input_string=result, pattern=pattern)
 
             # Если были найдены соответствующие строки
             if matched_lines:
                 for line in matched_lines:
                     # Поиск имени пакета в строке
-                    match = re.search(r'u0\s(.+?)/', line)
+                    match = re.search(r"u0\s(.+?)/", line)
                     if match:
                         # Возвращаем найденное значение
                         package_name = match.group(1)
@@ -354,12 +353,12 @@ class Adb:
 
             # Если не удалось найти имя пакета, возвращаем None
             logger.error("get_current_app_package() > None")
-            return None
-        except subprocess.CalledProcessError as e:
-            logger.error(f"{inspect.currentframe().f_code.co_name} > None")
+            return ""
+        except subprocess.CalledProcessError:
+            logger.error(f"{get_current_func_name()} > None")
             traceback_info = "".join(traceback.format_tb(sys.exc_info()[2]))
             logger.error(f"{sys.exc_info()[0]}\n{traceback_info}{sys.exc_info()[1]}")
-            return None
+            return ""
 
     @staticmethod
     def close_app(package: str) -> bool:
@@ -376,13 +375,13 @@ class Adb:
         """
         logger.info(f"close_app() < {package=}")
 
-        command = ['adb', 'shell', 'am', 'force-stop', package]
+        command = ["adb", "shell", "am", "force-stop", package]
         try:
-            subprocess.run(command, check=True)
+            subprocess.run(command, check=True)  # noqa: S603
             logger.info("close_app() > True")
             return True
-        except subprocess.CalledProcessError as e:
-            logger.error(f"{inspect.currentframe().f_code.co_name} > None")
+        except subprocess.CalledProcessError:
+            logger.error(f"{get_current_func_name()} > None")
             traceback_info = "".join(traceback.format_tb(sys.exc_info()[2]))
             logger.error(f"{sys.exc_info()[0]}\n{traceback_info}{sys.exc_info()[1]}")
             return False
@@ -427,13 +426,13 @@ class Adb:
         """
         logger.info("press_home()")
 
-        command = ['adb', 'shell', 'input', 'keyevent', 'KEYCODE_HOME']
+        command = ["adb", "shell", "input", "keyevent", "KEYCODE_HOME"]
         try:
-            subprocess.run(command, check=True)
+            subprocess.run(command, check=True)  # noqa: S603
             logger.info("press_home() > True")
             return True
-        except subprocess.CalledProcessError as e:
-            logger.error(f"{inspect.currentframe().f_code.co_name} > None")
+        except subprocess.CalledProcessError:
+            logger.error(f"{get_current_func_name()} > None")
             traceback_info = "".join(traceback.format_tb(sys.exc_info()[2]))
             logger.error(f"{sys.exc_info()[0]}\n{traceback_info}{sys.exc_info()[1]}")
             return False
@@ -449,13 +448,13 @@ class Adb:
         """
         logger.info("press_back()")
 
-        command = ['adb', 'shell', 'input', 'keyevent', 'KEYCODE_BACK']
+        command = ["adb", "shell", "input", "keyevent", "KEYCODE_BACK"]
         try:
-            subprocess.run(command, check=True)
+            subprocess.run(command, check=True)  # noqa: S603
             logger.info("press_back() > True")
             return True
-        except subprocess.CalledProcessError as e:
-            logger.error(f"{inspect.currentframe().f_code.co_name} > None")
+        except subprocess.CalledProcessError:
+            logger.error(f"{get_current_func_name()} > None")
             traceback_info = "".join(traceback.format_tb(sys.exc_info()[2]))
             logger.error(f"{sys.exc_info()[0]}\n{traceback_info}{sys.exc_info()[1]}")
             return False
@@ -471,13 +470,13 @@ class Adb:
         """
         logger.info("press_menu()")
 
-        command = ['adb', 'shell', 'input', 'keyevent', 'KEYCODE_MENU']
+        command = ["adb", "shell", "input", "keyevent", "KEYCODE_MENU"]
         try:
-            subprocess.run(command, check=True)
+            subprocess.run(command, check=True)  # noqa: S603
             logger.info("press_menu() > True")
             return True
-        except subprocess.CalledProcessError as e:
-            logger.error(f"{inspect.currentframe().f_code.co_name} > None")
+        except subprocess.CalledProcessError:
+            logger.error(f"{get_current_func_name()} > None")
             traceback_info = "".join(traceback.format_tb(sys.exc_info()[2]))
             logger.error(f"{sys.exc_info()[0]}\n{traceback_info}{sys.exc_info()[1]}")
             return False
@@ -497,13 +496,13 @@ class Adb:
         """
         logger.info(f"input_keycode_num_() < {num=}")
 
-        command = ['adb', 'shell', 'input', 'keyevent', f'KEYCODE_NUMPAD_{num}']
+        command = ["adb", "shell", "input", "keyevent", f"KEYCODE_NUMPAD_{num}"]
         try:
-            subprocess.run(command, check=True)
+            subprocess.run(command, check=True)  # noqa: S603
             logger.info("input_keycode_num_() > True")
             return True
-        except subprocess.CalledProcessError as e:
-            logger.error(f"{inspect.currentframe().f_code.co_name} > None")
+        except subprocess.CalledProcessError:
+            logger.error(f"{get_current_func_name()} > None")
             traceback_info = "".join(traceback.format_tb(sys.exc_info()[2]))
             logger.error(f"{sys.exc_info()[0]}\n{traceback_info}{sys.exc_info()[1]}")
             return False
@@ -523,41 +522,13 @@ class Adb:
         """
         logger.info(f"input_keycode() < {keycode=}")
 
-        command = ['adb', 'shell', 'input', 'keyevent', f'{keycode}']
+        command = ["adb", "shell", "input", "keyevent", f"{keycode}"]
         try:
-            subprocess.run(command, check=True)
+            subprocess.run(command, check=True)  # noqa: S603
             logger.info("input_keycode() > True")
             return True
-        except subprocess.CalledProcessError as e:
-            logger.error(f"{inspect.currentframe().f_code.co_name} > None")
-            traceback_info = "".join(traceback.format_tb(sys.exc_info()[2]))
-            logger.error(f"{sys.exc_info()[0]}\n{traceback_info}{sys.exc_info()[1]}")
-            return False
-
-    @staticmethod
-    def input_by_virtual_keyboard(text: str, keyboard: Dict[str, tuple]) -> bool:
-        """
-        Input text using a virtual keyboard by tapping on the corresponding coordinates for each character.
-
-        Args:
-            text : str
-                The text to be inputted.
-            keyboard : Dict[str, tuple]
-                A dictionary mapping each character to its corresponding coordinates on the virtual keyboard.
-
-        Returns:
-            bool
-                True if the input was successfully executed, False otherwise.
-        """
-        logger.info(f"input_by_virtual_keyboard() < {text=}, {keyboard=}")
-        try:
-            for char in text:
-                # Вызываем функцию tap с координатами, соответствующими символу char
-                Adb.tap(*keyboard[char])
-            logger.info("input_by_virtual_keyboard() > True")
-            return True
-        except subprocess.CalledProcessError as e:
-            logger.error(f"{inspect.currentframe().f_code.co_name} > None")
+        except subprocess.CalledProcessError:
+            logger.error(f"{get_current_func_name()} > None")
             traceback_info = "".join(traceback.format_tb(sys.exc_info()[2]))
             logger.error(f"{sys.exc_info()[0]}\n{traceback_info}{sys.exc_info()[1]}")
             return False
@@ -578,20 +549,20 @@ class Adb:
         logger.info(f"input_text() < {text=}")
 
         # Формируем команду для ввода текста с использованием ADB
-        command = ['adb', 'shell', 'input', 'text', text]
+        command = ["adb", "shell", "input", "text", text]
         try:
             # Выполняем команду
-            subprocess.run(command, check=True)
+            subprocess.run(command, check=True)  # noqa: S603
             logger.info("input_text() > True")
             return True
-        except subprocess.CalledProcessError as e:
-            logger.error(f"{inspect.currentframe().f_code.co_name} > None")
+        except subprocess.CalledProcessError:
+            logger.error(f"{get_current_func_name()} > None")
             traceback_info = "".join(traceback.format_tb(sys.exc_info()[2]))
             logger.error(f"{sys.exc_info()[0]}\n{traceback_info}{sys.exc_info()[1]}")
             return False
 
     @staticmethod
-    def tap(x: Union[str, int], y: Union[str, int]) -> bool:
+    def tap(x: str | int, y: str | int) -> bool:
         """
         Simulate a tap at the specified screen coordinates on the device using ADB.
 
@@ -608,20 +579,20 @@ class Adb:
         logger.info(f"tap() < {x=}, {y=}")
 
         # Формируем команду для выполнения нажатия по указанным координатам с использованием ADB
-        command = ['adb', 'shell', 'input', 'tap', str(x), str(y)]
+        command = ["adb", "shell", "input", "tap", str(x), str(y)]
         try:
-            subprocess.run(command, check=True)
+            subprocess.run(command, check=True)  # noqa: S603
             logger.info("tap() > True")
             return True
-        except subprocess.CalledProcessError as e:
-            logger.error(f"{inspect.currentframe().f_code.co_name} > None")
+        except subprocess.CalledProcessError:
+            logger.error(f"{get_current_func_name()} > None")
             traceback_info = "".join(traceback.format_tb(sys.exc_info()[2]))
             logger.error(f"{sys.exc_info()[0]}\n{traceback_info}{sys.exc_info()[1]}")
             return False
 
     @staticmethod
-    def swipe(start_x: Union[str, int], start_y: Union[str, int],
-              end_x: Union[str, int], end_y: Union[str, int],
+    def swipe(start_x: str | int, start_y: str | int,
+              end_x: str | int, end_y: str | int,
               duration: int = 300) -> bool:
         """
         Simulate a swipe gesture from the starting coordinates to the ending coordinates on the device using ADB.
@@ -645,20 +616,20 @@ class Adb:
         logger.info(f"swipe() < {start_x=}, {start_y=}, {end_x=}, {end_y=}, {duration=}")
 
         # Формируем команду для выполнения свайпа с использованием ADB
-        command = ['adb', 'shell', 'input', 'swipe', str(start_x), str(start_y), str(end_x), str(end_y), str(duration)]
+        command = ["adb", "shell", "input", "swipe", str(start_x), str(start_y), str(end_x), str(end_y), str(duration)]
         try:
             # Выполняем команду
-            subprocess.run(command, check=True)
+            subprocess.run(command, check=True)  # noqa: S603
             logger.info("swipe() > True")
             return True
-        except subprocess.CalledProcessError as e:
-            logger.error(f"{inspect.currentframe().f_code.co_name} > None")
+        except subprocess.CalledProcessError:
+            logger.error(f"{get_current_func_name()} > None")
             traceback_info = "".join(traceback.format_tb(sys.exc_info()[2]))
             logger.error(f"{sys.exc_info()[0]}\n{traceback_info}{sys.exc_info()[1]}")
             return False
 
     @staticmethod
-    def check_vpn(ip_address: str = '') -> bool:
+    def check_vpn(ip_address: str = "") -> bool:
         """
         Check if a VPN connection is established with the specified IP address.
 
@@ -676,7 +647,7 @@ class Adb:
         command = "adb shell netstat"
         try:
             # Выполняем команду и получаем вывод
-            output = subprocess.run(command, shell=True, capture_output=True, text=True, check=True)
+            output = subprocess.run(command, shell=True, capture_output=True, text=True, check=True)  # noqa: S602
 
             # Поиск строки
             lines = output.stdout.split("\n")
@@ -686,8 +657,8 @@ class Adb:
                     return True
             logger.info("check_vpn() False")
             return False
-        except subprocess.CalledProcessError as e:
-            logger.error(f"{inspect.currentframe().f_code.co_name} > None")
+        except subprocess.CalledProcessError:
+            logger.error(f"{get_current_func_name()} > None")
             traceback_info = "".join(traceback.format_tb(sys.exc_info()[2]))
             logger.error(f"{sys.exc_info()[0]}\n{traceback_info}{sys.exc_info()[1]}")
             return False
@@ -702,16 +673,15 @@ class Adb:
                 True if the logcat process was successfully stopped, False otherwise.
         """
         logger.info("stop_logcat()")
-        if Adb.is_process_exist(name='logcat'):
-            if Adb.kill_all(name='logcat'):
-                logger.info("stop_logcat() > True")
-                return True
+        if Adb.is_process_exist(name="logcat") and Adb.kill_all(name="logcat"):
+            logger.info("stop_logcat() > True")
+            return True
         logger.error("stop_logcat() > False")
         logger.info("stop_logcat() [Запущенного процесса logcat не обнаружено]")
         return False
 
     @staticmethod
-    def is_process_exist(name) -> bool:
+    def is_process_exist(name: str) -> bool:
         """
         Check if a process with the specified name is currently running on the device.
 
@@ -724,16 +694,16 @@ class Adb:
                 True if the process is running, False otherwise.
         """
         logger.info(f"is_process_exist() < {name=}")
-        command = ['adb', 'shell', 'ps']
+        command = ["adb", "shell", "ps"]
         try:
-            processes = subprocess.check_output(command, shell=True).decode().strip()
-        except subprocess.CalledProcessError as e:
-            logger.error(f"{inspect.currentframe().f_code.co_name} > None")
+            processes = subprocess.check_output(command, shell=True).decode().strip()  # noqa: S602
+        except subprocess.CalledProcessError:
+            logger.error(f"{get_current_func_name()} > None")
             traceback_info = "".join(traceback.format_tb(sys.exc_info()[2]))
             logger.error(f"{sys.exc_info()[0]}\n{traceback_info}{sys.exc_info()[1]}")
             return False
         # Разделение вывода на строки и удаление пустых строк
-        lines = processes.strip().split('\n')
+        lines = processes.strip().split("\n")
         # Проход по каждой строке вывода, начиная с 2-й строки, игнорируя заголовки
         for line in lines[1:]:
             # Разделение строки на столбцы по пробелам
@@ -769,15 +739,15 @@ class Adb:
 
         command = f"{command} nohup > /dev/null 2>&1 &"
         try:
-            subprocess.Popen(command, stdout=subprocess.DEVNULL)  # не добавлять with
+            subprocess.Popen(command, stdout=subprocess.DEVNULL)  # noqa: S603  # не добавлять with
             if process != "":
                 time.sleep(1)
                 if not Adb.is_process_exist(name=process):
                     return False
             logger.info("run_background_process() > True")
             return True
-        except subprocess.CalledProcessError as e:
-            logger.error(f"{inspect.currentframe().f_code.co_name} > None")
+        except subprocess.CalledProcessError:
+            logger.error(f"{get_current_func_name()} > None")
             traceback_info = "".join(traceback.format_tb(sys.exc_info()[2]))
             logger.error(f"{sys.exc_info()[0]}\n{traceback_info}{sys.exc_info()[1]}")
             return False
@@ -794,20 +764,20 @@ class Adb:
         logger.info("reload_adb()")
 
         try:
-            command = ['adb', 'kill-server']
-            subprocess.run(command, check=True)
-        except subprocess.CalledProcessError as e:
-            logger.error(f"{inspect.currentframe().f_code.co_name} > None")
+            command = ["adb", "kill-server"]
+            subprocess.run(command, check=True)  # noqa: S603
+        except subprocess.CalledProcessError:
+            logger.error(f"{get_current_func_name()} > None")
             traceback_info = "".join(traceback.format_tb(sys.exc_info()[2]))
             logger.error(f"{sys.exc_info()[0]}\n{traceback_info}{sys.exc_info()[1]}")
             return False
         # Ожидаем некоторое время перед запуском adb-сервера
         time.sleep(3)
         try:
-            command = ['adb', 'start-server']
-            subprocess.run(command, check=True)
-        except subprocess.CalledProcessError as e:
-            logger.error(f"{inspect.currentframe().f_code.co_name} > None")
+            command = ["adb", "start-server"]
+            subprocess.run(command, check=True)  # noqa: S603
+        except subprocess.CalledProcessError:
+            logger.error(f"{get_current_func_name()} > None")
             traceback_info = "".join(traceback.format_tb(sys.exc_info()[2]))
             logger.error(f"{sys.exc_info()[0]}\n{traceback_info}{sys.exc_info()[1]}")
             return False
@@ -815,7 +785,7 @@ class Adb:
         return True
 
     @staticmethod
-    def know_pid(name: str) -> Union[int, None]:
+    def know_pid(name: str) -> int | None:
         """
         Retrieve the process ID (PID) of a running process with the specified name.
 
@@ -828,16 +798,16 @@ class Adb:
                 The PID of the process if found, None otherwise.
         """
         logger.info(f"know_pid() < {name=}")
-        command = ['adb', 'shell', 'ps']
+        command = ["adb", "shell", "ps"]
         try:
-            processes = subprocess.check_output(command, shell=True).decode().strip()
-        except subprocess.CalledProcessError as e:
-            logger.error(f"{inspect.currentframe().f_code.co_name} > None")
+            processes = subprocess.check_output(command, shell=True).decode().strip()  # noqa: S602
+        except subprocess.CalledProcessError:
+            logger.error(f"{get_current_func_name()} > None")
             traceback_info = "".join(traceback.format_tb(sys.exc_info()[2]))
             logger.error(f"{sys.exc_info()[0]}\n{traceback_info}{sys.exc_info()[1]}")
             return None
         # Разделение вывода на строки и удаление пустых строк
-        lines = processes.strip().split('\n')
+        lines = processes.strip().split("\n")
         # Проход по каждой строке вывода, начиная с 2-й строки, игнорируя заголовки
         for line in lines[1:]:
             # Разделение строки на столбцы по пробелам
@@ -856,7 +826,7 @@ class Adb:
         return None
 
     @staticmethod
-    def kill_by_pid(pid: Union[str, int]) -> bool:
+    def kill_by_pid(pid: str | int) -> bool:
         """
         Terminate a process with the specified PID using ADB.
 
@@ -870,11 +840,11 @@ class Adb:
         """
         logger.info(f"kill_by_pid() < {pid=}")
 
-        command = ['adb', 'shell', 'kill', '-s', 'SIGINT', str(pid)]
+        command = ["adb", "shell", "kill", "-s", "SIGINT", str(pid)]
         try:
-            subprocess.call(command)
-        except subprocess.CalledProcessError as e:
-            logger.error(f"{inspect.currentframe().f_code.co_name} > None")
+            subprocess.call(command)  # noqa: S603
+        except subprocess.CalledProcessError:
+            logger.error(f"{get_current_func_name()} > None")
             traceback_info = "".join(traceback.format_tb(sys.exc_info()[2]))
             logger.error(f"{sys.exc_info()[0]}\n{traceback_info}{sys.exc_info()[1]}")
             return False
@@ -896,11 +866,11 @@ class Adb:
         """
         logger.info(f"kill_by_name() < {name=}")
 
-        command = ['adb', 'shell', 'pkill', '-l', 'SIGINT', str(name)]
+        command = ["adb", "shell", "pkill", "-l", "SIGINT", str(name)]
         try:
-            subprocess.call(command)
-        except subprocess.CalledProcessError as e:
-            logger.error(f"{inspect.currentframe().f_code.co_name} > None")
+            subprocess.call(command)  # noqa: S603
+        except subprocess.CalledProcessError:
+            logger.error(f"{get_current_func_name()} > None")
             traceback_info = "".join(traceback.format_tb(sys.exc_info()[2]))
             logger.error(f"{sys.exc_info()[0]}\n{traceback_info}{sys.exc_info()[1]}")
             return False
@@ -922,11 +892,11 @@ class Adb:
         """
         logger.info(f"kill_all() < {name=}")
 
-        command = ['adb', 'shell', 'pkill', '-f', str(name)]
+        command = ["adb", "shell", "pkill", "-f", str(name)]
         try:
-            subprocess.run(command, check=True)
-        except subprocess.CalledProcessError as e:
-            logger.error(f"{inspect.currentframe().f_code.co_name} > None")
+            subprocess.run(command, check=True)  # noqa: S603
+        except subprocess.CalledProcessError:
+            logger.error(f"{get_current_func_name()} > None")
             traceback_info = "".join(traceback.format_tb(sys.exc_info()[2]))
             logger.error(f"{sys.exc_info()[0]}\n{traceback_info}{sys.exc_info()[1]}")
             return False
@@ -948,11 +918,11 @@ class Adb:
         """
         logger.info(f"delete_files_from_internal_storage() < {path=}")
 
-        command = ['adb', 'shell', 'rm', '-rf', f'{path}*']
+        command = ["adb", "shell", "rm", "-rf", f"{path}*"]
         try:
-            subprocess.run(command, check=True)
-        except subprocess.CalledProcessError as e:
-            logger.error(f"{inspect.currentframe().f_code.co_name} > None")
+            subprocess.run(command, check=True)  # noqa: S603
+        except subprocess.CalledProcessError:
+            logger.error(f"{get_current_func_name()} > None")
             traceback_info = "".join(traceback.format_tb(sys.exc_info()[2]))
             logger.error(f"{sys.exc_info()[0]}\n{traceback_info}{sys.exc_info()[1]}")
             return False
@@ -960,7 +930,7 @@ class Adb:
         return True
 
     @staticmethod
-    def pull_video(source: str = None, destination: str = ".", delete: bool = True) -> bool:
+    def pull_video(source: str, destination: str = ".", delete: bool = True) -> bool:
         """
         Pull videos from the specified source directory on the device to the destination directory on the local machine.
 
@@ -980,27 +950,27 @@ class Adb:
         logger.info(f"pull_video() < {destination=}")
 
         if not source:
-            source = '/sdcard/Movies/'
-        if source.endswith('/'):
+            source = "/sdcard/Movies/"
+        if source.endswith("/"):
             source = source + "/"
-        if destination.endswith('/'):
+        if destination.endswith("/"):
             destination = destination + "/"
 
-        command = ['adb', 'pull', f'{source}', f'{destination}']
+        command = ["adb", "pull", f"{source}", f"{destination}"]
         try:
-            subprocess.run(command, check=True)
-        except subprocess.CalledProcessError as e:
-            logger.error(f"{inspect.currentframe().f_code.co_name} > None")
+            subprocess.run(command, check=True)  # noqa: S603
+        except subprocess.CalledProcessError:
+            logger.error(f"{get_current_func_name()} > None")
             traceback_info = "".join(traceback.format_tb(sys.exc_info()[2]))
             logger.error(f"{sys.exc_info()[0]}\n{traceback_info}{sys.exc_info()[1]}")
             return False
 
         if delete:
-            command = ['adb', 'shell', 'rm', '-rf', f'{source}*']
+            command = ["adb", "shell", "rm", "-rf", f"{source}*"]
             try:
-                subprocess.run(command, check=True)
-            except subprocess.CalledProcessError as e:
-                logger.error(f"{inspect.currentframe().f_code.co_name} > None")
+                subprocess.run(command, check=True)  # noqa: S603
+            except subprocess.CalledProcessError:
+                logger.error(f"{get_current_func_name()} > None")
                 traceback_info = "".join(traceback.format_tb(sys.exc_info()[2]))
                 logger.error(f"{sys.exc_info()[0]}\n{traceback_info}{sys.exc_info()[1]}")
                 return False
@@ -1019,11 +989,11 @@ class Adb:
         """
         logger.info("stop_video()")
 
-        command = ['adb', 'shell', 'pkill', '-l', 'SIGINT', 'screenrecord']
+        command = ["adb", "shell", "pkill", "-l", "SIGINT", "screenrecord"]
         try:
-            subprocess.call(command)
-        except subprocess.CalledProcessError as e:
-            logger.error(f"{inspect.currentframe().f_code.co_name} > None")
+            subprocess.call(command)  # noqa: S603
+        except subprocess.CalledProcessError:
+            logger.error(f"{get_current_func_name()} > None")
             traceback_info = "".join(traceback.format_tb(sys.exc_info()[2]))
             logger.error(f"{sys.exc_info()[0]}\n{traceback_info}{sys.exc_info()[1]}")
             return False
@@ -1031,8 +1001,8 @@ class Adb:
         return True
 
     @staticmethod
-    def record_video(path: str = "sdcard/Movies/", filename: str = "screenrecord.mp4") -> \
-            Union[subprocess.Popen[bytes], subprocess.Popen[Union[Union[str, bytes], Any]]]:
+    def record_video(path: str = "sdcard/Movies/", filename: str = "screenrecord.mp4") -> subprocess.Popen[
+                                                                                              bytes] | None:
         """
         Start recording a video on the device using ADB.
 
@@ -1047,17 +1017,17 @@ class Adb:
                 The Popen object representing the running video recording process if successful, None otherwise.
         """
         logger.info(f"record_video() < {filename}")
-        if path.endswith('/'):
+        if path.endswith("/"):
             path = path[:-1]
-        if filename.endswith('.mp4'):
+        if filename.endswith(".mp4"):
             filename = filename + ".mp4"
 
-        command = ['adb', 'shell', 'screenrecord', f'{path}/{filename}']
+        command = ["adb", "shell", "screenrecord", f"{path}/{filename}"]
         try:
             # Запускаем команду adb shell screenrecord для начала записи видео
-            return subprocess.Popen(command)
-        except subprocess.CalledProcessError as e:
-            logger.error(f"{inspect.currentframe().f_code.co_name} > None")
+            return subprocess.Popen(command)  # noqa: S603
+        except subprocess.CalledProcessError:
+            logger.error(f"{get_current_func_name()} > None")
             traceback_info = "".join(traceback.format_tb(sys.exc_info()[2]))
             logger.error(f"{sys.exc_info()[0]}\n{traceback_info}{sys.exc_info()[1]}")
             return None
@@ -1077,18 +1047,18 @@ class Adb:
             bool
                 True if the video recording was successfully started, False otherwise.
         """
-        if path.endswith('/'):
+        if path.endswith("/"):
             path = path[:-1]
-        if not filename.endswith('.mp4'):
+        if not filename.endswith(".mp4"):
             filename = filename + ".mp4"
 
-        command = ['adb', 'shell', 'screenrecord', f'{path}/{filename}']
+        command = ["adb", "shell", "screenrecord", f"{path}/{filename}"]
         try:
             # Запускаем команду adb shell screenrecord для начала записи видео
-            subprocess.Popen(command)  # не добавлять with
+            subprocess.Popen(command)  # noqa: S603  # не добавлять with
             return True
         except subprocess.CalledProcessError:
-            logger.error(f"{inspect.currentframe().f_code.co_name} > None")
+            logger.error(f"{get_current_func_name()} > None")
             traceback_info = "".join(traceback.format_tb(sys.exc_info()[2]))
             logger.error(f"{sys.exc_info()[0]}\n{traceback_info}{sys.exc_info()[1]}")
             return False
@@ -1104,11 +1074,11 @@ class Adb:
         """
         logger.info("reboot()")
 
-        command = ['adb', 'shell', 'reboot']
+        command = ["adb", "shell", "reboot"]
         try:
-            subprocess.call(command)
-        except subprocess.CalledProcessError as e:
-            logger.error(f"{inspect.currentframe().f_code.co_name} > None")
+            subprocess.call(command)  # noqa: S603
+        except subprocess.CalledProcessError:
+            logger.error(f"{get_current_func_name()} > None")
             traceback_info = "".join(traceback.format_tb(sys.exc_info()[2]))
             logger.error(f"{sys.exc_info()[0]}\n{traceback_info}{sys.exc_info()[1]}")
             return False
@@ -1116,7 +1086,7 @@ class Adb:
         return True
 
     @staticmethod
-    def get_screen_resolution() -> Union[Tuple[int, int], None]:
+    def get_screen_resolution() -> tuple[int, int] | None:
         """
         Retrieve the screen resolution of the connected device.
 
@@ -1126,22 +1096,22 @@ class Adb:
         """
         logger.info("get_screen_resolution()")
 
-        command = ['adb', 'shell', 'wm', 'size']
+        command = ["adb", "shell", "wm", "size"]
         try:
-            output = subprocess.check_output(command).decode()
+            output = subprocess.check_output(command).decode()  # noqa: S603
             if "Physical size" in output:
                 resolution_str = output.split(":")[1].strip()
                 width, height = resolution_str.split("x")
                 logger.info(f"get_screen_resolution() > {width=}, {height=}")
                 return int(width), int(height)
             logger.error(f"Unexpected output from adb: {output}")
-        except (subprocess.CalledProcessError, ValueError) as e:
-            logger.error(f"{inspect.currentframe().f_code.co_name} > None")
+        except (subprocess.CalledProcessError, ValueError):
+            logger.error(f"{get_current_func_name()} > None")
             traceback_info = "".join(traceback.format_tb(sys.exc_info()[2]))
             logger.error(f"{sys.exc_info()[0]}\n{traceback_info}{sys.exc_info()[1]}")
         return None
 
-    def get_packages_list(self) -> list:
+    def get_packages_list(self) -> list[str]:
         """
         Retrieve a list of all installed packages on the device.
 
@@ -1151,10 +1121,9 @@ class Adb:
         """
         packages_raw = self.execute(command="shell pm list packages")
         # Используем регулярное выражение для удаления "package:" из каждой строки
-        packages_raw = re.sub(r'package:', '', packages_raw)
+        packages_raw = re.sub(r"package:", "", packages_raw)
         # Разбиваем строки на список и удаляем пустые элементы
-        packages_list = [package.strip() for package in packages_raw.split('\n') if package.strip()]
-        return packages_list
+        return [package.strip() for package in packages_raw.split("\n") if package.strip()]
 
     @staticmethod
     def execute(command: str):
@@ -1170,5 +1139,5 @@ class Adb:
                 The output of the executed command as a string.
         """
         logger.info(f"execute() < {command}")
-        execute_command = ['adb', *command.split()]
-        return subprocess.check_output(execute_command).decode()
+        execute_command = ["adb", *command.split()]
+        return subprocess.check_output(execute_command).decode()  # noqa: S603
