@@ -1,5 +1,7 @@
 # shadowstep/terminal/terminal.py
+from __future__ import annotations
 
+import base64
 import logging
 import os
 import re
@@ -7,21 +9,25 @@ import subprocess
 import sys
 import time
 import traceback
-import inspect
-import base64
-from typing import Any, Union, Tuple
+from typing import TYPE_CHECKING, Any
 
 from appium.webdriver.webdriver import WebDriver
-from selenium.common import NoSuchDriverException, InvalidSessionIdException
+from selenium.common import InvalidSessionIdException, NoSuchDriverException
+
+from shadowstep.utils.utils import get_current_func_name
 
 # Configure the root logger (basic configuration)
-logging.basicConfig(level=logging.INFO, format='%(asctime)s - %(name)s - %(levelname)s - %(message)s')
+logging.basicConfig(level=logging.INFO, format="%(asctime)s - %(name)s - %(levelname)s - %(message)s")
 logger = logging.getLogger(__name__)
 
+if TYPE_CHECKING:
+    from shadowstep.base import ShadowstepBase
+    from shadowstep.terminal import Transport
 
-class NotProvideCredentials(Exception):
-    def __init__(self, message='Not provided credentials for ssh connection '
-                               'in connect() method (ssh_username, ssh_password)'):
+
+class NotProvideCredentialsError(Exception):
+    def __init__(self, message: str = "Not provided credentials for ssh connection "
+                                      "in connect() method (ssh_username, ssh_password)"):
         super().__init__(message)
         self.message = message
 
@@ -31,13 +37,13 @@ class Terminal:
     Allows you to perform adb actions using the appium server. Useful for remote connections
     Required ssh
     """
-    base: "Shadowstep" = None
-    transport: "Transport" = None
-    driver: WebDriver = None
+    base: ShadowstepBase
+    transport: Transport
+    driver: WebDriver
 
-    def __init__(self, base: "Shadowstep"):
-        self.base: "Shadowstep" = base
-        self.transport: "Transport" = base.transport
+    def __init__(self, base: ShadowstepBase):
+        self.base: ShadowstepBase = base
+        self.transport: Transport = base.transport
         self.driver: WebDriver = base.driver
 
     def __del__(self):
@@ -55,7 +61,7 @@ class Terminal:
         """
         for _ in range(tries):
             try:
-                return self.driver.execute_script("mobile: shell", {'command': command, 'args': [args]})
+                return self.driver.execute_script("mobile: shell", {"command": command, "args": [args]})
             except NoSuchDriverException:
                 self.base.reconnect()
             except InvalidSessionIdException:
@@ -64,6 +70,7 @@ class Terminal:
                 logger.error(e)
                 traceback_info = "".join(traceback.format_tb(sys.exc_info()[2]))
                 logger.error(traceback_info)
+        return None
 
     def push(self, source_path: str, remote_server_path: str, filename: str, destination: str, udid: str) -> bool:
         """
@@ -76,28 +83,28 @@ class Terminal:
         :param udid: The unique device identifier of the mobile device.
         :return: True if the file was successfully pushed, False otherwise.
         """
-        if self.transport is None:
-            raise NotProvideCredentials
         try:
             source_file_path = os.path.join(source_path, filename)
             remote_file_path = os.path.join(remote_server_path, filename)
             destination_file_path = f"{destination}/{filename}"
             self.transport.scp.put(files=source_file_path, remote_path=remote_file_path)
-            stdin, stdout, stderr = self.transport.ssh.exec_command(
-                f'adb -s {udid} push {remote_file_path} {destination_file_path}')
+            _, stdout, _ = self.transport.ssh.exec_command(
+                f"adb -s {udid} push {remote_file_path} {destination_file_path}")
             stdout_exit_status = stdout.channel.recv_exit_status()
             lines = stdout.readlines()
-            output = ''.join(lines)
+            output = "".join(lines)
             if stdout_exit_status != 0:
-                logger.error(f"{inspect.currentframe().f_code.co_name} {output=}")
+                logger.error(f"{get_current_func_name()} {output=}")
                 return False
-            logger.debug(f"{inspect.currentframe().f_code.co_name} {output=}")
+            logger.debug(f"{get_current_func_name()} {output=}")
             return True
         except NoSuchDriverException:
             self.base.reconnect()
+            return False
         except InvalidSessionIdException:
             self.base.reconnect()
-        except IOError as e:
+            return False
+        except OSError as e:
             logger.error("push()")
             logger.error(e)
             traceback_info = "".join(traceback.format_tb(sys.exc_info()[2]))
@@ -120,19 +127,21 @@ class Terminal:
                 # Если путь не указан, сохраняем в текущей директории
                 destination = os.path.join(os.getcwd(), os.path.basename(source))
 
-            file_contents_base64 = self.driver.assert_extension_exists('mobile: pullFile'). \
-                execute_script('mobile: pullFile', {'remotePath': source})
+            file_contents_base64 = self.driver.assert_extension_exists("mobile: pullFile"). \
+                execute_script("mobile: pullFile", {"remotePath": source})
             if not file_contents_base64:
                 return False
             decoded_contents = base64.b64decode(file_contents_base64)
-            with open(destination, 'wb') as file:
+            with open(destination, "wb") as file:
                 file.write(decoded_contents)
             return True
         except NoSuchDriverException:
             self.base.reconnect()
+            return False
         except InvalidSessionIdException:
             self.base.reconnect()
-        except IOError as e:
+            return False
+        except OSError as e:
             logger.error("appium_extended_terminal.pull")
             logger.error(e)
             traceback_info = "".join(traceback.format_tb(sys.exc_info()[2]))
@@ -158,7 +167,7 @@ class Terminal:
             logger.error(traceback_info)
             return False
 
-    def get_current_app_package(self) -> Union[str, None]:
+    def get_current_app_package(self) -> str:
         """
         Retrieves the package name of the currently focused application on the device.
 
@@ -166,19 +175,19 @@ class Terminal:
         """
         try:
             result = self.adb_shell(command="dumpsys", args="window windows")
-            lines = result.split('\n')
+            lines = result.split("\n")
             for line in lines:
-                if 'mCurrentFocus' in line or 'mFocusedApp' in line:
-                    matches = re.search(r'(([A-Za-z]{1}[A-Za-z\d_]*\.)+([A-Za-z][A-Za-z\d_]*)/)', line)
+                if "mCurrentFocus" in line or "mFocusedApp" in line:
+                    matches = re.search(r"(([A-Za-z]{1}[A-Za-z\d_]*\.)+([A-Za-z][A-Za-z\d_]*)/)", line)
                     if matches:
                         return matches.group(1)[:-1]  # removing trailing slash
-            return None
+            return ""
         except KeyError as e:
             logger.error("appium_extended_terminal.get_current_app_package()")
             logger.error(e)
             traceback_info = "".join(traceback.format_tb(sys.exc_info()[2]))
             logger.error(traceback_info)
-            return None
+            return ""
 
     def close_app(self, package: str) -> bool:
         """
@@ -208,9 +217,7 @@ class Terminal:
         """
         if not self.close_app(package=package):
             return False
-        if not self.start_activity(package=package, activity=activity):
-            return False
-        return True
+        return self.start_activity(package=package, activity=activity)
 
     def install_app(self, source: str, remote_server_path: str, filename: str, udid: str) -> bool:
         """
@@ -221,32 +228,30 @@ class Terminal:
         :param filename: The name of the application file.
         :param udid: The unique device identifier (UDID) of the target mobile device.
         :return: True if the application was successfully installed, False otherwise.
-        :raises NotProvideCredentials: If the transport credentials are not provided.
+        :raises NotProvideCredentialsError: If the transport credentials are not provided.
         """
-        if self.transport is None:
-            raise NotProvideCredentials
         try:
             source_filepath = os.path.join(source, filename)
             destination_filepath = os.path.join(remote_server_path, filename)
             self.transport.scp.put(files=source_filepath, remote_path=destination_filepath)
-            stdin, stdout, stderr = self.transport.ssh.exec_command(
-                f'adb -s {udid} install -r {destination_filepath}')
+            _, stdout, _ = self.transport.ssh.exec_command(
+                f"adb -s {udid} install -r {destination_filepath}")
             stdout_exit_status = stdout.channel.recv_exit_status()
             lines = stdout.readlines()
-            output = ''.join(lines)
+            output = "".join(lines)
             if stdout_exit_status != 0:
-                logger.error(f"{inspect.currentframe().f_code.co_name} {output=}")
+                logger.error(f"{get_current_func_name()} {output=}")
                 return False
-            logger.debug(f"{inspect.currentframe().f_code.co_name} {output=}")
+            logger.debug(f"{get_current_func_name()} {output=}")
             return True
-        except IOError as e:
+        except OSError as e:
             logger.error("appium_extended_terminal.push()")
             logger.error(e)
             traceback_info = "".join(traceback.format_tb(sys.exc_info()[2]))
             logger.error(traceback_info)
             return False
 
-    def is_app_installed(self, package) -> bool:
+    def is_app_installed(self, package: str) -> bool:
         """
         Checks if the specified application package is installed on the device.
 
@@ -257,7 +262,7 @@ class Terminal:
 
         try:
             result = self.adb_shell(command="pm", args="list packages")
-            if any([line.strip().endswith(package) for line in result.splitlines()]):
+            if any(line.strip().endswith(package) for line in result.splitlines()):
                 logger.debug("is_app_installed() > True")
                 return True
             logger.debug("is_app_installed() > False")
@@ -281,8 +286,10 @@ class Terminal:
             return True
         except NoSuchDriverException:
             self.base.reconnect()
+            return False
         except InvalidSessionIdException:
             self.base.reconnect()
+            return False
         except KeyError as e:
             logger.error("appium_extended_terminal.uninstall_app()")
             logger.error(e)
@@ -410,8 +417,8 @@ class Terminal:
             logger.error(traceback_info)
             return False
 
-    def swipe(self, start_x: Union[str, int], start_y: Union[str, int],
-              end_x: Union[str, int], end_y: Union[str, int], duration: int = 300) -> bool:
+    def swipe(self, start_x: str | int, start_y: str | int,
+              end_x: str | int, end_y: str | int, duration: int = 300) -> bool:
         """
         Simulates a swipe gesture from one point to another on the device's screen.
 
@@ -503,7 +510,7 @@ class Terminal:
                           end_y=height // 2,
                           duration=duration)
 
-    def check_vpn(self, ip_address: str = '') -> bool:
+    def check_vpn(self, ip_address: str = "") -> bool:
         """
         Checks if a VPN connection is established on the device.
 
@@ -512,7 +519,7 @@ class Terminal:
         """
         try:
             output = self.adb_shell(command="netstat", args="")
-            lines = output.split('\n')
+            lines = output.split("\n")
             for line in lines:
                 if ip_address in line and "ESTABLISHED" in line:
                     logger.debug("check_VPN() True")
@@ -553,7 +560,7 @@ class Terminal:
                     return False
         return True
 
-    def know_pid(self, name: str) -> Union[int, None]:
+    def know_pid(self, name: str) -> int | None:
         """
         Retrieves the process ID (PID) of the specified process name.
 
@@ -564,7 +571,7 @@ class Terminal:
         if name not in processes:
             logger.error("know_pid() [Процесс не обнаружен]")
             return None
-        lines = processes.strip().split('\n')
+        lines = processes.strip().split("\n")
         for line in lines[1:]:
             columns = line.split()
             if len(columns) >= 9:
@@ -586,7 +593,7 @@ class Terminal:
         if name not in processes:
             logger.debug("is_process_exist() > False")
             return False
-        lines = processes.strip().split('\n')
+        lines = processes.strip().split("\n")
         for line in lines[1:]:
             columns = line.split()
             if len(columns) >= 9:
@@ -673,7 +680,7 @@ class Terminal:
             return False
         return True
 
-    def delete_files_from_internal_storage(self, path) -> bool:
+    def delete_files_from_internal_storage(self, path: str) -> bool:
         """
         Deletes files from the internal storage of the device.
 
@@ -699,7 +706,7 @@ class Terminal:
         :return: True if the file was successfully deleted, False otherwise.
         """
         try:
-            if path.endswith('/'):
+            if path.endswith("/"):
                 path = path[:-1]
             self.adb_shell(command="rm", args=f"-rf {path}/{filename}")
         except KeyError as e:
@@ -731,7 +738,7 @@ class Terminal:
             return False
         return True
 
-    def stop_video(self, **options: Any) -> Union[bytes, None]:
+    def stop_video(self, **options: Any) -> bytes | None:
         """
         Stops the video recording of the device screen and returns the recorded video data (Base64 bytes).
 
@@ -757,13 +764,13 @@ class Terminal:
         Reboots the device safely. If adb connection drops, ignores the error.
         """
         try:
-            self.adb_shell(command='reboot')
+            self.adb_shell(command="reboot")
             return True
         except Exception as e:
             logger.warning(f"Reboot likely initiated. Caught exception: {e}")
             return True
 
-    def get_screen_resolution(self) -> Union[Tuple[int, int], None]:
+    def get_screen_resolution(self) -> tuple[int, int]:
         """
         Retrieves the screen resolution of the device.
 
@@ -771,16 +778,18 @@ class Terminal:
                  or None if the resolution couldn't be retrieved.
         """
         try:
-            output = self.adb_shell(command='wm', args='size')
+            output = self.adb_shell(command="wm", args="size")
             if "Physical size" in output:
                 resolution_str = output.split(":")[1].strip()
                 width, height = resolution_str.split("x")
                 return int(width), int(height)
-        except KeyError as e:
-            logger.error("appium_extended_terminal.get_screen_resolution")
+            logger.warning(f"{get_current_func_name()}: Physical size not in output")
+            return 0, 0
+        except Exception as e:
             logger.error(e)
             traceback_info = "".join(traceback.format_tb(sys.exc_info()[2]))
             logger.error(traceback_info)
+            raise
 
     def past_text(self, text: str, tries: int = 3) -> None:
         """
@@ -789,14 +798,14 @@ class Terminal:
         for _ in range(tries):
             try:
                 self.driver.set_clipboard_text(text=text)
-                self.input_keycode('279')
+                self.input_keycode("279")
                 return
             except NoSuchDriverException:
                 self.base.reconnect()
             except InvalidSessionIdException:
                 self.base.reconnect()
 
-    def get_prop(self) -> dict:
+    def get_prop(self) -> dict[str, Any]:
         """
         Retrieves system properties from the device.
 
@@ -813,7 +822,6 @@ class Terminal:
                 result_dict[key] = value
             except ValueError:
                 continue
-
         return result_dict
 
     def get_prop_hardware(self) -> str:
@@ -822,7 +830,7 @@ class Terminal:
 
         :return: A string representing the hardware information.
         """
-        return self.get_prop()['ro.boot.hardware']
+        return self.get_prop()["ro.boot.hardware"]
 
     def get_prop_model(self) -> str:
         """
@@ -830,7 +838,7 @@ class Terminal:
 
         :return: A string representing the model name of the device.
         """
-        return self.get_prop()['ro.product.model']
+        return self.get_prop()["ro.product.model"]
 
     def get_prop_serial(self) -> str:
         """
@@ -838,7 +846,7 @@ class Terminal:
 
         :return: A string representing the serial number of the device.
         """
-        return self.get_prop()['ro.serialno']
+        return self.get_prop()["ro.serialno"]
 
     def get_prop_build(self) -> str:
         """
@@ -846,7 +854,7 @@ class Terminal:
 
         :return: A string representing the build description of the device.
         """
-        return self.get_prop()['ro.build.description']
+        return self.get_prop()["ro.build.description"]
 
     def get_prop_device(self) -> str:
         """
@@ -854,7 +862,7 @@ class Terminal:
 
         :return: A string representing the device name.
         """
-        return self.get_prop()['ro.product.device']
+        return self.get_prop()["ro.product.device"]
 
     def get_prop_uin(self) -> str:
         """
@@ -862,18 +870,17 @@ class Terminal:
 
         :return: A string representing the unique identification number.
         """
-        return self.get_prop()['sys.atol.uin']
+        return self.get_prop()["sys.atol.uin"]
 
-    def get_packages(self) -> list:
+    def get_packages(self) -> list[str]:
         """
         Retrieves the list of installed packages on the device.
 
         :return: A list of package names.
         """
-        output = self.adb_shell(command='pm', args='list packages')
-        lines = output.strip().split('\n')
-        packages = [line.split(':')[-1].replace('\r', '') for line in lines]
-        return packages
+        output = self.adb_shell(command="pm", args="list packages")
+        lines = output.strip().split("\n")
+        return [line.split(":")[-1].replace("\r", "") for line in lines]
 
     def get_package_path(self, package: str) -> str:
         """
@@ -882,25 +889,25 @@ class Terminal:
         :param package: The name of the package.
         :return: The path to the APK file.
         """
-        return self.adb_shell(command='pm', args=f'path {package}'). \
-            replace('package:', ''). \
-            replace('\r', ''). \
-            replace('\n', '')
+        return self.adb_shell(command="pm", args=f"path {package}"). \
+            replace("package:", ""). \
+            replace("\r", ""). \
+            replace("\n", "")
 
-    def pull_package(self, package: str, path: str = '', filename: str = 'temp.apk'):
+    def pull_package(self, package: str, path: str = "", filename: str = "temp._apk"):
         """
         Pulls the APK file of the specified package from the device to the local machine.
 
         :param package: The package name of the app.
         :param path: The local path where the APK file will be saved. Default is current directory.
-        :param filename: The name of the APK file. If not provided, a default name 'temp.apk' will be used.
+        :param filename: The name of the APK file. If not provided, a default name 'temp._apk' will be used.
         """
         package_path = self.get_package_path(package=package)
-        if not filename.endswith('.apk'):
-            filename = f"{filename}.apk"
+        if not filename.endswith("._apk"):
+            filename = f"{filename}._apk"
         self.pull(source=package_path, destination=os.path.join(path, filename))
 
-    def get_package_manifest(self, package: str) -> dict:
+    def get_package_manifest(self, package: str) -> dict[str, Any]:
         """
         Retrieves the manifest of the specified package from the device.
 
@@ -908,28 +915,28 @@ class Terminal:
         :return: A dictionary representing the package manifest.
         """
         if not os.path.exists("test"):
-            os.makedirs(name='test')
+            os.makedirs(name="test")
 
         self.pull_package(package=package, path="test",
-                          filename="temp.apk")
+                          filename="temp._apk")
 
-        command = ["aapt", "dump", "badging", os.path.join("test", "temp.apk")]
+        command = ["aapt", "dump", "badging", os.path.join("test", "temp._apk")]
         try:
-            output: str = str(subprocess.check_output(command)).strip()
-        except subprocess.CalledProcessError as error:
+            output: str = str(subprocess.check_output(command)).strip()  # noqa: S603
+        except subprocess.CalledProcessError:
             return {}
-        output = output.replace('\\r\\n', ' ').replace('b"', '').replace('"', '').replace(":'", ": '")
+        output = output.replace("\\r\\n", " ").replace('b"', "").replace('"', "").replace(":'", ": '")
         list_of_elements = output.split()
         result = {}
         current_key = None
 
         for element in list_of_elements:
-            if element.endswith(':'):
+            if element.endswith(":"):
                 result[element] = []
                 current_key = element
                 continue
             result[current_key].append(element.replace("'", ""))
 
-        os.remove(os.path.join('test', 'temp.apk'))
+        os.remove(os.path.join("test", "temp._apk"))
 
         return result
