@@ -1,7 +1,8 @@
 # shadowstep/element/should.py
 
-from typing import Any, Optional, Tuple
-from .element import Element
+from typing import Any
+
+from shadowstep.element.element import Element
 
 
 class Should:
@@ -18,9 +19,9 @@ class Should:
         """Delegates unknown attribute access to the underlying Element instance."""
         try:
             return getattr(self.element, name)
-        except AttributeError:
+        except AttributeError as error:
             raise AttributeError(
-                f"'Should' has no attribute '{name}', and '{self.element.__class__.__name__}' also does not have it.")
+                f"'Should' has no attribute '{name}', and '{self.element.__class__.__name__}' also does not have it.") from error
 
 
 class _ShouldBase:
@@ -32,9 +33,11 @@ class _ShouldBase:
 
     def _assert(self, condition: bool, message: str) -> None:
         if self.negate:
-            assert not condition, "[should.not] " + message
+            if condition:
+                raise AssertionError("[should.not] " + message)
         else:
-            assert condition, "[should] " + message
+            if not condition:
+                raise AssertionError("[should] " + message)
 
 
 class _ShouldHave(_ShouldBase):
@@ -65,7 +68,7 @@ class _ShouldHave(_ShouldBase):
         self._assert(actual == expected, f"have.package: expected '{expected}', got '{actual}'")
         return Should(self.element)
 
-    def bounds(self, expected: Tuple[int, int, int, int]) -> Should:
+    def bounds(self, expected: str) -> Should:
         actual = self.element.get_attribute("bounds")
         self._assert(actual == expected, f"have.bounds: expected {expected}, got {actual}")
         return Should(self.element)
