@@ -1,9 +1,11 @@
-import inspect
+from __future__ import annotations
+
 import logging
 import textwrap
 from pathlib import Path
+from typing import Any
 
-from icecream import ic
+from shadowstep.utils.utils import get_current_func_name
 
 
 class PageObjectMerger:
@@ -12,28 +14,23 @@ class PageObjectMerger:
 
     def merge(self, file1: str, file2: str, output_path: str) -> str:
         """
-        Производит слияние
+        merge pages
         """
-        self.logger.info(f"{inspect.currentframe().f_code.co_name}")
+        self.logger.info(f"{get_current_func_name()}")
         page1 = self.parse(file1)
         page2 = self.parse(file2)
         imports = self.get_imports(page1)
         class_name = self.get_class_name(page1)
         methods1 = self.get_methods(page1)
         methods2 = self.get_methods(page2)
-        # self.logger.info(f"{page1=}")
-        # self.logger.info(f"{imports=}")
-        # self.logger.info(f"{class_name=}")
-        # self.logger.info(f"{methods1=}")
         unique_methods = self.remove_duplicates(methods1, methods2)
-        # self.logger.info(f"{unique_methods=}")
         self.write_to_file(filepath=output_path,
                            imports=imports,
                            class_name=class_name,
                            unique_methods=unique_methods)
         return output_path
 
-    def parse(self, file) -> str:
+    def parse(self, file: str | Path) -> str:
         """
         Reads and returns the full content of a Python file as a UTF-8 string.
 
@@ -43,12 +40,11 @@ class PageObjectMerger:
         Returns:
             str: Raw content of the file.
         """
-        self.logger.debug(f"{inspect.currentframe().f_code.co_name}")
+        self.logger.debug(f"{get_current_func_name()}")
         try:
-            with open(file, "r", encoding="utf-8") as f:
-                content = f.read()
+            with open(file, encoding="utf-8") as f:
                 # self.logger.info(f"{content=}")
-                return content
+                return f.read()
         except Exception as e:
             self.logger.error(f"Failed to read {file}: {e}")
             raise
@@ -63,9 +59,9 @@ class PageObjectMerger:
         Returns:
             str: All import lines joined by newline.
         """
-        self.logger.debug(f"{inspect.currentframe().f_code.co_name}")
+        self.logger.debug(f"{get_current_func_name()}")
         lines = page.splitlines()
-        import_lines = []
+        import_lines: list[str] = []
         for line in lines:
             stripped = line.strip()
             if stripped.startswith("import ") or stripped.startswith("from "):
@@ -90,7 +86,7 @@ class PageObjectMerger:
         Raises:
             ValueError: Если не найдено определение класса.
         """
-        self.logger.info(f"{inspect.currentframe().f_code.co_name}")
+        self.logger.info(f"{get_current_func_name()}")
         for line in page.splitlines():
             stripped = line.strip()
             self.logger.info(f"{stripped=}")
@@ -99,7 +95,7 @@ class PageObjectMerger:
                 return line.rstrip()
         raise ValueError("No class definition found in the given source.")
 
-    def get_methods(self, page: str) -> dict:
+    def get_methods(self, page: str) -> dict[str, Any]:
         """
         Извлекает методы и property-блоки через \n\n-разделение, с нормализацией отступов.
 
@@ -109,7 +105,7 @@ class PageObjectMerger:
         Returns:
             dict: имя_метода -> текст_метода
         """
-        self.logger.debug(f"{inspect.currentframe().f_code.co_name}")
+        self.logger.debug(f"{get_current_func_name()}")
 
         methods = {}
         blocks = page.split("\n\n")
@@ -118,7 +114,8 @@ class PageObjectMerger:
             block = textwrap.dedent(block)  # <<< ВАЖНО: УБИРАЕМ ЛИШНИЕ ВЛОЖЕННОСТИ
             stripped = block.strip()
 
-            if not stripped.startswith("def ") and not stripped.startswith("@property"):
+            if not stripped.startswith("def ") and not stripped.startswith("@property") and not stripped.startswith(
+                    "@current_page"):
                 continue
 
             lines = block.splitlines()
@@ -140,37 +137,15 @@ class PageObjectMerger:
 
         return methods
 
-    def remove_duplicates(self, methods1: dict, methods2: dict) -> dict:
-        """
-        Удаляет дубликаты методов (по имени и тексту), объединяя оба словаря.
-        При конфликте (одинаковое имя, разный код) оставляет версию из methods1.
-
-        Args:
-            methods1 (dict): Метод-словарь из первого файла.
-            methods2 (dict): Метод-словарь из второго файла.
-
-        Returns:
-            dict: Объединённый словарь уникальных методов.
-        """
-        self.logger.debug(f"{inspect.currentframe().f_code.co_name}")
+    def remove_duplicates(self, methods1: dict[str, Any], methods2: dict[str, Any]) -> dict[str, Any]:
+        self.logger.debug(f"{get_current_func_name()}")
 
         unique_methods = {}
 
-        # Сначала добавим всё из первого файла
         for name, body in methods1.items():
             unique_methods[name] = body
 
-        # Теперь добавляем из второго файла, если:
-        # - имя не встречалось
-        # - либо тело полностью совпадает (дубликат, но безопасный)
         for name, body in methods2.items():
-            # self.logger.info("===================================")
-            # self.logger.info(f"{name=}")
-            # self.logger.info(f"{body=}")
-            # self.logger.info(f"{unique_methods[name].strip()=}")
-            # self.logger.info(f"{body.strip()=}")
-            # self.logger.info(f"{unique_methods[name].strip() == body.strip()}")
-            # self.logger.info("===================================")
             if name not in unique_methods:
                 unique_methods[name] = body
             elif unique_methods[name].strip() == body.strip():
@@ -185,22 +160,12 @@ class PageObjectMerger:
             filepath: str,
             imports: str,
             class_name: str,
-            unique_methods: dict,
+            unique_methods: dict[str, Any],
             encoding: str = "utf-8"
     ) -> None:
-        self.logger.debug(f"{inspect.currentframe().f_code.co_name}")
-        lines = []
+        self.logger.debug(f"{get_current_func_name()}")
+        lines: list[str] = [imports.strip(), "", "", class_name.strip(), ""]
 
-        # Импорты
-        lines.append(imports.strip())
-        lines.append("")  # Пустая строка
-        lines.append("")  # Пустая строка
-
-        # Заголовок класса
-        lines.append(class_name.strip())
-        lines.append("")  # Пустая строка
-
-        # Методы
         for name, body in unique_methods.items():
             if name == "recycler" or name == "is_current_page":
                 continue
@@ -209,7 +174,7 @@ class PageObjectMerger:
             lines.append(method_lines)
             lines.append("")  # Пустая строка между методами
 
-        if "recycler" in unique_methods.keys():
+        if "recycler" in unique_methods:
             body = unique_methods["recycler"]
             clean_body = textwrap.dedent(body)  # убрать вложенные отступы
             method_lines = textwrap.indent(clean_body, "    ")  # вложить внутрь класса
@@ -226,4 +191,3 @@ class PageObjectMerger:
         output_path = Path(filepath)
         output_path.parent.mkdir(parents=True, exist_ok=True)
         output_path.write_text(content, encoding=encoding)
-
