@@ -4,6 +4,8 @@ import logging
 import os
 import subprocess
 
+from shadowstep.utils.utils import get_current_func_name
+
 logger = logging.getLogger(__name__)
 
 
@@ -15,13 +17,13 @@ class Aapt:
         Получает название пакета APK-файла с помощью команды aapt.
         Возвращает название пакета.
         """
-        logger.info(f"get_package_name() < {path_to_apk}")
+        logger.info(f"{get_current_func_name()} < {path_to_apk}")
 
         command = ["aapt", "dump", "badging", os.path.join(path_to_apk)]
 
         try:
             # Выполнение команды и получение вывода
-            output: str = str(subprocess.check_output(command)).strip()
+            output: str = str(subprocess.check_output(command)).strip()  # noqa: S603
 
             # Извлечение строки, содержащей информацию о пакете
             start_index = output.index("package: name='") + len("package: name='")
@@ -35,10 +37,10 @@ class Aapt:
             raise  # Выбрасываем исключение дальше
 
         except ValueError:
-            logger.error(f"Could not find package name in the output.")
+            logger.error("Could not find package name in the output.")
             raise  # Выбрасываем исключение дальше
 
-        logger.info(f"get_package_name() > {package_name}")
+        logger.info(f"{get_current_func_name()} > {package_name}")
         # Возвращение названия пакета в виде строки
         return package_name
 
@@ -48,13 +50,13 @@ class Aapt:
         Получает название запускаемой активности из APK-файла с помощью команды aapt.
         Возвращает название активности в виде строки.
         """
-        logger.info(f"get_launchable_activity_from_apk() < {path_to_apk}")
+        logger.info(f"{get_current_func_name()} < {path_to_apk}")
 
         command = ["aapt", "dump", "badging", path_to_apk]
 
         try:
             # Выполнение команды и получение вывода
-            output = subprocess.check_output(command, universal_newlines=True).strip()
+            output = subprocess.check_output(command, universal_newlines=True).strip()  # noqa: S603
 
             # Извлечение строки, содержащей информацию о запускаемой активности
             package_line = next(line for line in output.splitlines() if line.startswith("launchable-activity"))
@@ -63,7 +65,7 @@ class Aapt:
             launchable_activity = package_line.split("'")[1]
 
             # Возвращение названия активности в виде строки
-            logger.info(f"get_launchable_activity_from_apk() > {launchable_activity}")
+            logger.info(f"{get_current_func_name()} > {launchable_activity}")
             return launchable_activity
         except subprocess.CalledProcessError as e:
             logger.error(f"Could not extract launchable activity. Error: {str(e)}")
